Wrapper for bin_interactions() that interacts with the filesystem

Places putative interactions into equally-sized bins based on the distances they span, and stores these on disk.

bin_interactions_fs(
  interactions_file,
  fragments_file,
  output_dir,
  min_dist = 2500,
  max_dist = Inf,
  bins = 5
)

Arguments

interactions_file
Path to (tsv or csv) file containing putative interactions. Its first line must state the columns names: baitID, preyID, N. Subsequent lines report one putative interaction each: a bait fragment ID, prey fragment ID and readcount.

fragments_file
Path to (bed) file containing fragment information. Its first line must state the column names: chrom, chromStart, chromEnd, ID. Each subsequent line reports the chromosome, starting coordinate, ending coordinate and ID of a fragment.

output_dir
Directory where the generated bins will be stored. Will be created if it does not exist.

min_dist
Minimum distance interactions must span to be included in the analysis. Distance is defined between fragment midpoints.

max_dist
Maximum distance interactions can span to be included in the analysis. Distance is defined between fragment midpoints.

bins
Number of bins to place the interactions into.

Value

List containing the output directory and an overview of bins.

Examples

base = system.file("extdata",package="peaky")

interactions_file = paste0(base,"/counts.tsv")
fragments_file = paste0(base,"/fragments.bed")
bins_dir = paste0(base,"/bins")
# \donttest{
BI = bin_interactions_fs(interactions_file, fragments_file, output_dir=bins_dir)

#> 31-08-2020 16:48:04
#> Reading interactions from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/counts.tsv
#> 31-08-2020 16:48:04
#> Reading fragment information from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fragments.bed
#> 31-08-2020 16:48:04
#> Calculating fragment characteristics...
#> 31-08-2020 16:48:04
#> Adding fragment characteristics for baits...
#> 31-08-2020 16:48:04
#> Adding fragment characteristics for preys...
#> 31-08-2020 16:48:04
#> Calculating interaction distances...
#> 31-08-2020 16:48:04
#> Calculating total trans-chromosomal read counts for each bait...
#> 31-08-2020 16:48:04
#> Modelling those as a function of bait chromosome...
#> GAMLSS-RS iteration 1: Global Deviance = 363.9654 
#> GAMLSS-RS iteration 2: Global Deviance = 363.9654 
#> 31-08-2020 16:48:04
#> Adding trans-chromosomal interactivity covariate for preys that were also baited (0 for preys not baited)...
#> 31-08-2020 16:48:04
#> Excluding 2 interactions that are too proximal (distance < 2500 bp)...
#> 31-08-2020 16:48:04
#> Excluding 0 interactions that are too distal (distance > Inf bp)...
#> 31-08-2020 16:48:04
#> Assigning 5 distance bins...
#> 31-08-2020 16:48:04
#> Done.
#> 31-08-2020 16:48:04
#> Saving binned interactions:
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> 31-08-2020 16:48:05
#> Saving bin details to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bins.txt

print(BI)

#> $output_dir
#> [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins"
#> 
#> $interactions
#>        baitID preyID N b.chr     b.mid b.length p.chr     p.mid p.length
#>     1:  53559  52096 1     1 212337456     8519     1 207345336     1932
#>     2:  53559  52107 1     1 212337456     8519     1 207380122      891
#>     3:  53559  52112 1     1 212337456     8519     1 207394962     6067
#>     4:  53559  52113 1     1 212337456     8519     1 207398127      260
#>     5:  53559  52121 1     1 212337456     8519     1 207434420     1859
#>    ---                                                                  
#> 78050: 661722 830195 1     6 170834793     4398     X 155188600      619
#> 78051: 661722 831633 1     6 170834793     4398     Y   7238968     2825
#> 78052: 661722 832798 1     6 170834793     4398     Y  14832472     5436
#> 78053: 661722 833895 1     6 170834793     4398     Y  18220566     2150
#> 78054: 661722 835350 1     6 170834793     4398     Y  22944058     3103
#>            dist b.trans b.trans_res p.trans_res dist.bin
#>     1: -4992120     151  0.09918876           0        5
#>     2: -4957334     151  0.09918876           0        5
#>     3: -4942494     151  0.09918876           0        5
#>     4: -4939329     151  0.09918876           0        5
#>     5: -4903036     151  0.09918876           0        5
#>    ---                                                  
#> 78050:       NA    1719  2.21400352           0     <NA>
#> 78051:       NA    1719  2.21400352           0     <NA>
#> 78052:       NA    1719  2.21400352           0     <NA>
#> 78053:       NA    1719  2.21400352           0     <NA>
#> 78054:       NA    1719  2.21400352           0     <NA>
#> 
#> $bins
#>    dist.bin dist.abs.min dist.abs.max interactions
#> 1:        1         2907       733944         9064
#> 2:        2       734021      1590586         9063
#> 3:        3      1590733      2558841         9063
#> 4:        4      2559052      3664236         9063
#> 5:        5      3664323      4999963         9064
#> 6:     <NA>           NA           NA        32737
#> 

# }