Interprets models fitted with peaky

Extracts the marginal posterior probabilities of contact and other relevant statistics from the RJMCMC results and merges these with bait-associated information.

interpret_peaky(bait, peaks, omega_power, log_file = NA)

Arguments

bait
Data table containing the putative interactions of a bait, having the columns 'baitID', 'dist', and 'residual'. These report the bait ID, its distance to putative preys, and the adjusted readcounts for its interactions with them, respectively.

peaks
The models fitted by peaky.

omega_power
The same value as used when running peaky, i.e. the expected decay of adjusted read counts around a truly interacting prey. See details.

Value

A data table containing bait-associated information, posterior probabilities of contact and other statistics.

Details

The steepness of the function to be fitted to putative peaks is determined by \(\omega\) according to \(\beta \exp{- \abs{\omega * d}}\), where \(\beta\) represents peak height and \(d\) the distance from the center of the peak in bp.

Examples

base = system.file("extdata",package="peaky")
interactions_file = paste0(base,"/counts.tsv")
fragments_file = paste0(base,"/fragments.bed")
interactions = data.table::fread(interactions_file)
fragments = data.table::fread(fragments_file)

BI = bin_interactions(interactions, fragments, bins=5)

#> 31-08-2020 16:48:05
#> Calculating fragment characteristics...
#> 31-08-2020 16:48:05
#> Adding fragment characteristics for baits...
#> 31-08-2020 16:48:05
#> Adding fragment characteristics for preys...
#> 31-08-2020 16:48:05
#> Calculating interaction distances...
#> 31-08-2020 16:48:05
#> Calculating total trans-chromosomal read counts for each bait...
#> 31-08-2020 16:48:05
#> Modelling those as a function of bait chromosome...
#> GAMLSS-RS iteration 1: Global Deviance = 363.9654 
#> GAMLSS-RS iteration 2: Global Deviance = 363.9654 
#> 31-08-2020 16:48:05
#> Adding trans-chromosomal interactivity covariate for preys that were also baited (0 for preys not baited)...
#> 31-08-2020 16:48:05
#> Excluding 2 interactions that are too proximal (distance < 2500 bp)...
#> 31-08-2020 16:48:05
#> Excluding 0 interactions that are too distal (distance > Inf bp)...
#> 31-08-2020 16:48:05
#> Assigning 5 distance bins...
#> 31-08-2020 16:48:05
#> Done.

models = by(BI$interactions, BI$interactions$dist.bin, model_bin, subsample_size=1000)

#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 16:48:05
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 16:48:05
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 7626.698 
#> GAMLSS-RS iteration 2: Global Deviance = 7525.301 
#> GAMLSS-RS iteration 3: Global Deviance = 7498.109 
#> GAMLSS-RS iteration 4: Global Deviance = 7495.143 
#> GAMLSS-RS iteration 5: Global Deviance = 7494.994 
#> GAMLSS-RS iteration 6: Global Deviance = 7494.989 
#> 31-08-2020 16:48:05
#> Converged: YES
#> Iterations: 6
#> 
#> Coefficients:
#> (Intercept)	16.1950380153013
#>  log(abs(dist))	-1.05480805807603
#>  b.trans_res	0.201053864014002
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.0043482722907471
#>  sqrt(p.length)	-0.000797305953026929
#> 31-08-2020 16:48:05
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 16:48:05
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 16:48:05
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 4428.732 
#> GAMLSS-RS iteration 2: Global Deviance = 4426.55 
#> GAMLSS-RS iteration 3: Global Deviance = 4425.271 
#> GAMLSS-RS iteration 4: Global Deviance = 4424.465 
#> GAMLSS-RS iteration 5: Global Deviance = 4423.873 
#> GAMLSS-RS iteration 6: Global Deviance = 4423.447 
#> GAMLSS-RS iteration 7: Global Deviance = 4423.136 
#> GAMLSS-RS iteration 8: Global Deviance = 4422.931 
#> GAMLSS-RS iteration 9: Global Deviance = 4422.779 
#> GAMLSS-RS iteration 10: Global Deviance = 4422.684 
#> 31-08-2020 16:48:06
#> Converged: YES
#> Iterations: 10
#> 
#> Coefficients:
#> (Intercept)	31.4738576504188
#>  log(abs(dist))	-2.19391017768305
#>  b.trans_res	0.440085225877817
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00418257438205598
#>  sqrt(p.length)	-0.00245107217397082
#> 31-08-2020 16:48:06
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 16:48:06
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 16:48:06
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 2917.877 
#> GAMLSS-RS iteration 2: Global Deviance = 2895.991 
#> GAMLSS-RS iteration 3: Global Deviance = 2888.73 
#> GAMLSS-RS iteration 4: Global Deviance = 2885.239 
#> GAMLSS-RS iteration 5: Global Deviance = 2883.263 
#> GAMLSS-RS iteration 6: Global Deviance = 2882.021 
#> GAMLSS-RS iteration 7: Global Deviance = 2881.177 
#> GAMLSS-RS iteration 8: Global Deviance = 2880.575 
#> GAMLSS-RS iteration 9: Global Deviance = 2880.136 
#> GAMLSS-RS iteration 10: Global Deviance = 2879.802 
#> GAMLSS-RS iteration 11: Global Deviance = 2879.541 
#> GAMLSS-RS iteration 12: Global Deviance = 2879.325 
#> GAMLSS-RS iteration 13: Global Deviance = 2879.165 
#> GAMLSS-RS iteration 14: Global Deviance = 2879.019 
#> GAMLSS-RS iteration 15: Global Deviance = 2878.913 
#> GAMLSS-RS iteration 16: Global Deviance = 2878.828 
#> 31-08-2020 16:48:06
#> Converged: YES
#> Iterations: 16
#> 
#> Coefficients:
#> (Intercept)	39.1204453553899
#>  log(abs(dist))	-2.66288140414565
#>  b.trans_res	0.319451019194821
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00812367123742881
#>  sqrt(p.length)	-0.00353897364373813
#> 31-08-2020 16:48:06
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 16:48:06
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 16:48:06
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 2175.003 
#> GAMLSS-RS iteration 2: Global Deviance = 2162.55 
#> GAMLSS-RS iteration 3: Global Deviance = 2158.494 
#> GAMLSS-RS iteration 4: Global Deviance = 2156.664 
#> GAMLSS-RS iteration 5: Global Deviance = 2155.709 
#> GAMLSS-RS iteration 6: Global Deviance = 2155.154 
#> GAMLSS-RS iteration 7: Global Deviance = 2154.816 
#> GAMLSS-RS iteration 8: Global Deviance = 2154.599 
#> GAMLSS-RS iteration 9: Global Deviance = 2154.458 
#> GAMLSS-RS iteration 10: Global Deviance = 2154.361 
#> 31-08-2020 16:48:07
#> Converged: YES
#> Iterations: 10
#> 
#> Coefficients:
#> (Intercept)	10.6442849733832
#>  log(abs(dist))	-0.718792131367532
#>  b.trans_res	0.376659169752014
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00728731997101535
#>  sqrt(p.length)	0.00115720339699804
#> 31-08-2020 16:48:07
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 16:48:07
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 16:48:07
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 1545.331 
#> GAMLSS-RS iteration 2: Global Deviance = 1535.117 
#> GAMLSS-RS iteration 3: Global Deviance = 1530.387 
#> GAMLSS-RS iteration 4: Global Deviance = 1527.623 
#> GAMLSS-RS iteration 5: Global Deviance = 1525.804 
#> GAMLSS-RS iteration 6: Global Deviance = 1524.504 
#> GAMLSS-RS iteration 7: Global Deviance = 1523.538 
#> GAMLSS-RS iteration 8: Global Deviance = 1522.784 
#> GAMLSS-RS iteration 9: Global Deviance = 1522.178 
#> GAMLSS-RS iteration 10: Global Deviance = 1521.688 
#> GAMLSS-RS iteration 11: Global Deviance = 1521.276 
#> GAMLSS-RS iteration 12: Global Deviance = 1520.937 
#> GAMLSS-RS iteration 13: Global Deviance = 1520.646 
#> GAMLSS-RS iteration 14: Global Deviance = 1520.396 
#> GAMLSS-RS iteration 15: Global Deviance = 1520.178 
#> GAMLSS-RS iteration 16: Global Deviance = 1519.992 
#> GAMLSS-RS iteration 17: Global Deviance = 1519.827 
#> GAMLSS-RS iteration 18: Global Deviance = 1519.665 
#> GAMLSS-RS iteration 19: Global Deviance = 1519.519 
#> GAMLSS-RS iteration 20: Global Deviance = 1519.386 
#> GAMLSS-RS iteration 21: Global Deviance = 1519.263 
#> GAMLSS-RS iteration 22: Global Deviance = 1519.15 
#> GAMLSS-RS iteration 23: Global Deviance = 1519.065 
#> 31-08-2020 16:48:08
#> Converged: YES
#> Iterations: 23
#> 
#> Coefficients:
#> (Intercept)	69.3147184507793
#>  log(abs(dist))	-4.61596150051082
#>  b.trans_res	0.178320680821925
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00398565826072352
#>  sqrt(p.length)	0.00117263507783666
#> 31-08-2020 16:48:08
#> Obtaining normalized randomized quantile residuals for the full dataset...

residuals = lapply(models, "[[", "residuals")
bins = split(BI$interactions, BI$interactions$dist.bin)

BTS = split_baits(bins, residuals)

#> 31-08-2020 16:48:08
#> Calculating p-values and performing bin-wise FDR-adjustments...

relevant_bait = BTS[baitID==618421]
omega_power=4.7
PKS = peaky(relevant_bait, omega_power, iterations=1e6)

#> 31-08-2020 16:48:08
#> Constructing distance matrix with omega=10^4.7...
#> 1e+06 iterations will be run...
#> Setting user specified argument AlphaPriorMu 
#> Setting user specified argument AlphaPriorSd 
#> Setting user specified argument Alpha_Initial_Value 
#> Setting user specified argument GaussianResidual_Initial_Value 
#> Setting user specified argument GaussianResidualPrior_UnifArg1 
#> Setting user specified argument GaussianResidualPrior_UnifArg2 
#> Setting user specified argument GaussianResidualPriorFamily 
#> Setting user specified argument Add_Move_Probability 
#> Setting user specified argument Delete_Move_Probability 
#> Setting user specified argument Swap_Move_Probability 
#> 0 observations deleted due to missingnessWriting data into an input file for BGLiMS...
#> Data written in 0 hrs 0 mins and 59 seconds.
#> Calling BGLiMS Java algorithm with command: 
#> java -jar "/home/cq/R/x86_64-pc-linux-gnu-library/3.6/R2BGLiMS/BGLiMS/BGLiMS.jar" "/tmp/RtmpBhKk4I/file3c395e96b88a_Aug31164809_Arguments.txt" "/tmp/RtmpBhKk4I/file3c3943591c9d_Aug31164809_Data.txt" "/tmp/RtmpBhKk4I/file3c393588ea8d_Aug31164809_Results.txt" 1000000 0 100 100000 152572 1 1 0 -1 1 1 1 1707 
#> BGLiMS Java computation finished in 0 hrs 4 mins and 54 seconds.
#> Reading BGLiMS posterior output...
#> Posterior output processed in 0 hrs 0 mins and 7 seconds.
#> Finished.

interpret_peaky(relevant_bait, PKS, omega_power)

#> 31-08-2020 16:54:09
#> Calculating alphas...
#> 31-08-2020 16:54:09
#> Calculating posteriors...
#> 31-08-2020 16:54:10
#> Calculating betas means and medians...
#> 31-08-2020 16:54:11
#> Constructing distance matrix with omega=10^4.7...
#> Finding RJMCMC predictions based on mean betas...
#> 31-08-2020 16:54:11
#> Wrapping up...