Wrapper for interpret_peaky() that interacts with the filesystem

Extracts the marginal posterior probabilities of contact and other relevant statistics from the RJMCMC results and merges these with bait-associated information.

interpret_peaky_fs(rjmcmclist, index, baits_dir, output_dir, omega_power)

Arguments

rjmcmclist
Path to the list of RJMCMC results to be interpreted.

index
Which RJMCMC result on the list to process, with 1 corresponding to the first one on the list.

output_dir
Directory where the merged RJMCMC results and bait information will be stored, one file per bait. Will be created if it does not exist.

omega_power
The same value as used when running peaky, i.e. the expected decay of adjusted read counts around a truly interacting prey. See details.

Value

The output directory.

Details

The steepness of the function to be fitted to putative peaks is determined by \(\omega\) according to \(\beta \exp{- \abs{\omega * d}}\), where \(\beta\) represents peak height and \(d\) the distance from the center of the peak in bp.

Examples

base = system.file("extdata",package="peaky")
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

bin_interactions_fs(interactions_file, fragments_file, output_dir=bins_dir)

#> 31-08-2020 16:54:11
#> Reading interactions from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/counts.tsv
#> 31-08-2020 16:54:11
#> Reading fragment information from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fragments.bed
#> 31-08-2020 16:54:11
#> Calculating fragment characteristics...
#> 31-08-2020 16:54:11
#> Adding fragment characteristics for baits...
#> 31-08-2020 16:54:11
#> Adding fragment characteristics for preys...
#> 31-08-2020 16:54:11
#> Calculating interaction distances...
#> 31-08-2020 16:54:11
#> Calculating total trans-chromosomal read counts for each bait...
#> 31-08-2020 16:54:11
#> Modelling those as a function of bait chromosome...
#> GAMLSS-RS iteration 1: Global Deviance = 363.9654 
#> GAMLSS-RS iteration 2: Global Deviance = 363.9654 
#> 31-08-2020 16:54:11
#> Adding trans-chromosomal interactivity covariate for preys that were also baited (0 for preys not baited)...
#> 31-08-2020 16:54:11
#> Excluding 2 interactions that are too proximal (distance < 2500 bp)...
#> 31-08-2020 16:54:11
#> Excluding 0 interactions that are too distal (distance > Inf bp)...
#> 31-08-2020 16:54:11
#> Assigning 5 distance bins...
#> 31-08-2020 16:54:11
#> Done.
#> 31-08-2020 16:54:11
#> Saving binned interactions:
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> 31-08-2020 16:54:11
#> Saving bin details to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bins.txt

#> $output_dir
#> [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins"
#> 
#> $interactions
#>        baitID preyID N b.chr     b.mid b.length p.chr     p.mid p.length
#>     1:  53559  52096 1     1 212337456     8519     1 207345336     1932
#>     2:  53559  52107 1     1 212337456     8519     1 207380122      891
#>     3:  53559  52112 1     1 212337456     8519     1 207394962     6067
#>     4:  53559  52113 1     1 212337456     8519     1 207398127      260
#>     5:  53559  52121 1     1 212337456     8519     1 207434420     1859
#>    ---                                                                  
#> 78050: 661722 830195 1     6 170834793     4398     X 155188600      619
#> 78051: 661722 831633 1     6 170834793     4398     Y   7238968     2825
#> 78052: 661722 832798 1     6 170834793     4398     Y  14832472     5436
#> 78053: 661722 833895 1     6 170834793     4398     Y  18220566     2150
#> 78054: 661722 835350 1     6 170834793     4398     Y  22944058     3103
#>            dist b.trans b.trans_res p.trans_res dist.bin
#>     1: -4992120     151  0.09440699           0        5
#>     2: -4957334     151  0.09440699           0        5
#>     3: -4942494     151  0.09440699           0        5
#>     4: -4939329     151  0.09440699           0        5
#>     5: -4903036     151  0.09440699           0        5
#>    ---                                                  
#> 78050:       NA    1719  2.21285387           0     <NA>
#> 78051:       NA    1719  2.21285387           0     <NA>
#> 78052:       NA    1719  2.21285387           0     <NA>
#> 78053:       NA    1719  2.21285387           0     <NA>
#> 78054:       NA    1719  2.21285387           0     <NA>
#> 
#> $bins
#>    dist.bin dist.abs.min dist.abs.max interactions
#> 1:        1         2907       733944         9064
#> 2:        2       734021      1590586         9063
#> 3:        3      1590733      2558841         9063
#> 4:        4      2559052      3664236         9063
#> 5:        5      3664323      4999963         9064
#> 6:     <NA>           NA           NA        32737
#> 

fits_dir = paste0(base,"/fits")

for(bin_index in 1:5){
  model_bin_fs(bins_dir,bin_index,output_dir=fits_dir,subsample_size=1000)
}

#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 16:54:11
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 16:54:11
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 7629.385 
#> GAMLSS-RS iteration 2: Global Deviance = 7483.95 
#> GAMLSS-RS iteration 3: Global Deviance = 7408.069 
#> GAMLSS-RS iteration 4: Global Deviance = 7384.574 
#> GAMLSS-RS iteration 5: Global Deviance = 7381.903 
#> GAMLSS-RS iteration 6: Global Deviance = 7381.741 
#> GAMLSS-RS iteration 7: Global Deviance = 7381.733 
#> 31-08-2020 16:54:12
#> Converged: YES
#> Iterations: 7
#> 
#> Coefficients:
#> (Intercept)	16.388521079909
#>  log(abs(dist))	-1.09262424086401
#>  b.trans_res	0.173059612221204
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00193207581116291
#>  sqrt(p.length)	0.000139431485553219
#> 31-08-2020 16:54:12
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_1.rds...
#> 31-08-2020 16:54:12
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_1.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 16:54:12
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 16:54:12
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 4148.525 
#> GAMLSS-RS iteration 2: Global Deviance = 4147.294 
#> GAMLSS-RS iteration 3: Global Deviance = 4146.707 
#> GAMLSS-RS iteration 4: Global Deviance = 4146.381 
#> GAMLSS-RS iteration 5: Global Deviance = 4146.182 
#> GAMLSS-RS iteration 6: Global Deviance = 4146.048 
#> GAMLSS-RS iteration 7: Global Deviance = 4145.96 
#> 31-08-2020 16:54:12
#> Converged: YES
#> Iterations: 7
#> 
#> Coefficients:
#> (Intercept)	21.3265272674249
#>  log(abs(dist))	-1.47568917080874
#>  b.trans_res	0.417832665714363
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00536714950792311
#>  sqrt(p.length)	0.00117188141725302
#> 31-08-2020 16:54:12
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_2.rds...
#> 31-08-2020 16:54:12
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_2.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 16:54:12
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 16:54:12
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 2804.944 
#> GAMLSS-RS iteration 2: Global Deviance = 2791.854 
#> GAMLSS-RS iteration 3: Global Deviance = 2787.646 
#> GAMLSS-RS iteration 4: Global Deviance = 2785.751 
#> GAMLSS-RS iteration 5: Global Deviance = 2784.756 
#> GAMLSS-RS iteration 6: Global Deviance = 2784.182 
#> GAMLSS-RS iteration 7: Global Deviance = 2783.83 
#> GAMLSS-RS iteration 8: Global Deviance = 2783.601 
#> GAMLSS-RS iteration 9: Global Deviance = 2783.451 
#> GAMLSS-RS iteration 10: Global Deviance = 2783.347 
#> GAMLSS-RS iteration 11: Global Deviance = 2783.273 
#> 31-08-2020 16:54:13
#> Converged: YES
#> Iterations: 11
#> 
#> Coefficients:
#> (Intercept)	41.6661909738447
#>  log(abs(dist))	-2.84726227333898
#>  b.trans_res	0.33070445216004
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00748639332530992
#>  sqrt(p.length)	0.0011598376847124
#> 31-08-2020 16:54:13
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_3.rds...
#> 31-08-2020 16:54:13
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_3.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 16:54:13
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 16:54:13
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 2201.881 
#> GAMLSS-RS iteration 2: Global Deviance = 2189.511 
#> GAMLSS-RS iteration 3: Global Deviance = 2185.841 
#> GAMLSS-RS iteration 4: Global Deviance = 2184.265 
#> GAMLSS-RS iteration 5: Global Deviance = 2183.469 
#> GAMLSS-RS iteration 6: Global Deviance = 2183.026 
#> GAMLSS-RS iteration 7: Global Deviance = 2182.766 
#> GAMLSS-RS iteration 8: Global Deviance = 2182.609 
#> GAMLSS-RS iteration 9: Global Deviance = 2182.509 
#> GAMLSS-RS iteration 10: Global Deviance = 2182.443 
#> 31-08-2020 16:54:13
#> Converged: YES
#> Iterations: 10
#> 
#> Coefficients:
#> (Intercept)	10.4410089377557
#>  log(abs(dist))	-0.683171754532988
#>  b.trans_res	0.318350601684629
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00931591964983597
#>  sqrt(p.length)	-0.00192930849040938
#> 31-08-2020 16:54:13
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_4.rds...
#> 31-08-2020 16:54:14
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_4.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 16:54:14
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 16:54:14
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 1547.297 
#> GAMLSS-RS iteration 2: Global Deviance = 1535.577 
#> GAMLSS-RS iteration 3: Global Deviance = 1529.943 
#> GAMLSS-RS iteration 4: Global Deviance = 1526.585 
#> GAMLSS-RS iteration 5: Global Deviance = 1524.356 
#> GAMLSS-RS iteration 6: Global Deviance = 1522.753 
#> GAMLSS-RS iteration 7: Global Deviance = 1521.552 
#> GAMLSS-RS iteration 8: Global Deviance = 1520.608 
#> GAMLSS-RS iteration 9: Global Deviance = 1519.857 
#> GAMLSS-RS iteration 10: Global Deviance = 1519.249 
#> GAMLSS-RS iteration 11: Global Deviance = 1518.742 
#> GAMLSS-RS iteration 12: Global Deviance = 1518.315 
#> GAMLSS-RS iteration 13: Global Deviance = 1517.93 
#> GAMLSS-RS iteration 14: Global Deviance = 1517.592 
#> GAMLSS-RS iteration 15: Global Deviance = 1517.293 
#> GAMLSS-RS iteration 16: Global Deviance = 1517.054 
#> GAMLSS-RS iteration 17: Global Deviance = 1516.849 
#> GAMLSS-RS iteration 18: Global Deviance = 1516.636 
#> GAMLSS-RS iteration 19: Global Deviance = 1516.473 
#> GAMLSS-RS iteration 20: Global Deviance = 1516.33 
#> GAMLSS-RS iteration 21: Global Deviance = 1516.158 
#> GAMLSS-RS iteration 22: Global Deviance = 1516.001 
#> GAMLSS-RS iteration 23: Global Deviance = 1515.91 
#> 31-08-2020 16:54:15
#> Converged: YES
#> Iterations: 23
#> 
#> Coefficients:
#> (Intercept)	89.0939677610244
#>  log(abs(dist))	-5.91193341591523
#>  b.trans_res	0.25852908819546
#>  p.trans_res	NA
#>  sqrt(b.length)	0.00206808248489218
#>  sqrt(p.length)	-0.00798302869632052
#> 31-08-2020 16:54:15
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_5.rds...
#> 31-08-2020 16:54:15
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_5.rds...

baits_dir = paste0(base,"/baits")

split_baits_fs(bins_dir,residuals_dir = fits_dir, indices=1:5, output_dir = baits_dir)

#> 31-08-2020 16:54:15
#> Loading bins...
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> 31-08-2020 16:54:15
#> Loading residuals...
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_1.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_2.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_3.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_4.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_5.rds
#> 31-08-2020 16:54:15
#> Calculating p-values and performing bin-wise FDR-adjustments...
#> 31-08-2020 16:54:15
#> Saving baits...
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_53559.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_71638.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_95958.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_96180.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_102201.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119435.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119522.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_144385.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_162008.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_175487.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_203500.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_243669.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_262362.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_268979.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_272961.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_312780.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_328019.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_341315.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_363262.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_366588.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_466090.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_582536.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_618421.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_658681.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_661722.rds
#> 31-08-2020 16:54:15
#> Plotting baits...

#> 31-08-2020 16:54:17
#> Done.

#> $output_dir
#> [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits"
#> 
#> $bait_paths
#>  [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_53559.rds" 
#>  [2] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_71638.rds" 
#>  [3] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_95958.rds" 
#>  [4] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_96180.rds" 
#>  [5] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_102201.rds"
#>  [6] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119435.rds"
#>  [7] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119522.rds"
#>  [8] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_144385.rds"
#>  [9] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_162008.rds"
#> [10] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_175487.rds"
#> [11] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_203500.rds"
#> [12] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_243669.rds"
#> [13] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_262362.rds"
#> [14] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_268979.rds"
#> [15] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_272961.rds"
#> [16] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_312780.rds"
#> [17] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_328019.rds"
#> [18] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_341315.rds"
#> [19] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_363262.rds"
#> [20] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_366588.rds"
#> [21] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_466090.rds"
#> [22] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_582536.rds"
#> [23] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_618421.rds"
#> [24] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_658681.rds"
#> [25] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_661722.rds"
#> 

baitlist = paste0(baits_dir,"/baitlist.txt")
rjmcmc_dir = paste0(base,"/rjmcmc")
omega_power = 4.7
# \donttest{
for(i in 1:3){
 peaky_fs(baitlist,i,output_dir=rjmcmc_dir,omega_power=omega_power)
}

#> 31-08-2020 16:54:17
#> Loaded bait 53559 from path: /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_53559.rds...
#> 31-08-2020 16:54:17
#> Constructing distance matrix with omega=10^4.7...
#> 1e+06 iterations will be run...
#> Setting user specified argument AlphaPriorMu 
#> Setting user specified argument AlphaPriorSd 
#> Setting user specified argument Alpha_Initial_Value 
#> Setting user specified argument GaussianResidual_Initial_Value 
#> Setting user specified argument GaussianResidualPrior_UnifArg1 
#> Setting user specified argument GaussianResidualPrior_UnifArg2 
#> Setting user specified argument GaussianResidualPriorFamily 
#> Setting user specified argument Add_Move_Probability 
#> Setting user specified argument Delete_Move_Probability 
#> Setting user specified argument Swap_Move_Probability 
#> 0 observations deleted due to missingnessWriting data into an input file for BGLiMS...
#> Data written in 0 hrs 0 mins and 6 seconds.
#> Calling BGLiMS Java algorithm with command: 
#> java -jar "/home/cq/R/x86_64-pc-linux-gnu-library/3.6/R2BGLiMS/BGLiMS/BGLiMS.jar" "/tmp/RtmpBhKk4I/file3c3910f300e0_Aug31165417_Arguments.txt" "/tmp/RtmpBhKk4I/file3c39680da635_Aug31165417_Data.txt" "/tmp/RtmpBhKk4I/file3c392960c1d3_Aug31165417_Results.txt" 1000000 0 100 100000 46714 1 1 0 -1 1 1 1 826 
#> BGLiMS Java computation finished in 0 hrs 1 mins and 57 seconds.
#> Reading BGLiMS posterior output...
#> Posterior output processed in 0 hrs 0 mins and 2 seconds.
#> Finished.
#> 31-08-2020 16:56:23
#> Saving RJMCMC results to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc/rjmcmc_53559.rds
#> 31-08-2020 16:56:24
#> Done.
#> 31-08-2020 16:56:24
#> Loaded bait 71638 from path: /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_71638.rds...
#> 31-08-2020 16:56:24
#> Constructing distance matrix with omega=10^4.7...
#> 1e+06 iterations will be run...
#> Setting user specified argument AlphaPriorMu 
#> Setting user specified argument AlphaPriorSd 
#> Setting user specified argument Alpha_Initial_Value 
#> Setting user specified argument GaussianResidual_Initial_Value 
#> Setting user specified argument GaussianResidualPrior_UnifArg1 
#> Setting user specified argument GaussianResidualPrior_UnifArg2 
#> Setting user specified argument GaussianResidualPriorFamily 
#> Setting user specified argument Add_Move_Probability 
#> Setting user specified argument Delete_Move_Probability 
#> Setting user specified argument Swap_Move_Probability 
#> 0 observations deleted due to missingnessWriting data into an input file for BGLiMS...
#> Data written in 0 hrs 0 mins and 59 seconds.
#> Calling BGLiMS Java algorithm with command: 
#> java -jar "/home/cq/R/x86_64-pc-linux-gnu-library/3.6/R2BGLiMS/BGLiMS/BGLiMS.jar" "/tmp/RtmpBhKk4I/file3c396f89b96a_Aug31165624_Arguments.txt" "/tmp/RtmpBhKk4I/file3c392b66c2d2_Aug31165624_Data.txt" "/tmp/RtmpBhKk4I/file3c395ee9ac60_Aug31165624_Results.txt" 1000000 0 100 100000 830263 1 1 0 -1 1 1 1 2014 
#> BGLiMS Java computation finished in 0 hrs 5 mins and 39 seconds.
#> Reading BGLiMS posterior output...
#> Posterior output processed in 0 hrs 0 mins and 7 seconds.
#> Finished.
#> 31-08-2020 17:03:09
#> Saving RJMCMC results to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc/rjmcmc_71638.rds
#> 31-08-2020 17:03:10
#> Done.
#> 31-08-2020 17:03:10
#> Loaded bait 95958 from path: /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_95958.rds...
#> 31-08-2020 17:03:10
#> Constructing distance matrix with omega=10^4.7...
#> 1e+06 iterations will be run...
#> Setting user specified argument AlphaPriorMu 
#> Setting user specified argument AlphaPriorSd 
#> Setting user specified argument Alpha_Initial_Value 
#> Setting user specified argument GaussianResidual_Initial_Value 
#> Setting user specified argument GaussianResidualPrior_UnifArg1 
#> Setting user specified argument GaussianResidualPrior_UnifArg2 
#> Setting user specified argument GaussianResidualPriorFamily 
#> Setting user specified argument Add_Move_Probability 
#> Setting user specified argument Delete_Move_Probability 
#> Setting user specified argument Swap_Move_Probability 
#> 0 observations deleted due to missingnessWriting data into an input file for BGLiMS...
#> Data written in 0 hrs 1 mins and 0 seconds.
#> Calling BGLiMS Java algorithm with command: 
#> java -jar "/home/cq/R/x86_64-pc-linux-gnu-library/3.6/R2BGLiMS/BGLiMS/BGLiMS.jar" "/tmp/RtmpBhKk4I/file3c392b6d7e5b_Aug31170311_Arguments.txt" "/tmp/RtmpBhKk4I/file3c39f1d8058_Aug31170311_Data.txt" "/tmp/RtmpBhKk4I/file3c39270e2418_Aug31170311_Results.txt" 1000000 0 100 100000 145038 1 1 0 -1 1 1 1 2037 
#> BGLiMS Java computation finished in 0 hrs 5 mins and 45 seconds.
#> Reading BGLiMS posterior output...
#> Posterior output processed in 0 hrs 0 mins and 7 seconds.
#> Finished.
#> 31-08-2020 17:10:03
#> Saving RJMCMC results to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc/rjmcmc_95958.rds
#> 31-08-2020 17:10:04
#> Done.

rjmcmc_list(rjmcmc_dir)

#> 31-08-2020 17:10:04
#> Looking for RJMCMC results in /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc
#> Found 6 files in total.
#> Found 3 files containing RJMCMC results.

#> $filename
#> [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc/rjmcmclist.txt"
#> 
#> $rjmcmc_paths
#> [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc/rjmcmc_53559.rds"
#> [2] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc/rjmcmc_71638.rds"
#> [3] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc/rjmcmc_95958.rds"
#> 

rjmcmclist = paste0(rjmcmc_dir,"/rjmcmclist.txt")
baits_rjmcmc_dir = paste0(base,"/baits_rjmcmc")
interpret_peaky_fs(rjmcmclist,1,baits_dir,baits_rjmcmc_dir,omega_power)

#> Loading RJMCMC result and bait:
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc/rjmcmc_53559.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_53559.rds
#> 31-08-2020 17:10:04
#> Calculating alphas...
#> 31-08-2020 17:10:04
#> Calculating posteriors...
#> 31-08-2020 17:10:05
#> Calculating betas means and medians...
#> 31-08-2020 17:10:05
#> Constructing distance matrix with omega=10^4.7...
#> Finding RJMCMC predictions based on mean betas...
#> 31-08-2020 17:10:05
#> Wrapping up...
#> 31-08-2020 17:10:05
#> Saving the bait with its RJMCMC results to/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits_rjmcmc/bait_rjmcmc_53559.rds

#> $output_path
#> [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits_rjmcmc/bait_rjmcmc_53559.rds"
#> 
#> $output
#>      baitID preyID N b.chr     b.mid b.length p.chr     p.mid p.length     dist
#>   1:  53559  52096 1     1 212337456     8519     1 207345336     1932 -4992120
#>   2:  53559  52107 1     1 212337456     8519     1 207380122      891 -4957334
#>   3:  53559  52112 1     1 212337456     8519     1 207394962     6067 -4942494
#>   4:  53559  52113 1     1 212337456     8519     1 207398127      260 -4939329
#>   5:  53559  52121 1     1 212337456     8519     1 207434420     1859 -4903036
#>  ---                                                                           
#> 822:  53559  55157 1     1 212337456     8519     1 217191554     2183  4854098
#> 823:  53559  55184 1     1 212337456     8519     1 217272896     2230  4935440
#> 824:  53559  55185 1     1 212337456     8519     1 217275236     2449  4937780
#> 825:  53559  55193 1     1 212337456     8519     1 217299256     3869  4961800
#> 826:  53559  55197 1     1 212337456     8519     1 217318317     1404  4980861
#>      b.trans b.trans_res p.trans_res dist.bin   residual       p.res   fdr.res
#>   1:     151  0.09440699           0        5 -1.8029810 0.071391195 0.9706424
#>   2:     151  0.09440699           0        5  0.6397962 0.522305081 0.9990728
#>   3:     151  0.09440699           0        5  0.3856788 0.699734586 0.9990728
#>   4:     151  0.09440699           0        5  0.3764401 0.706589731 0.9990728
#>   5:     151  0.09440699           0        5 -2.7977548 0.005145916 0.7169123
#>  ---                                                                          
#> 822:     151  0.09440699           0        5  0.4274242 0.669070378 0.9990728
#> 823:     151  0.09440699           0        5  0.4353834 0.663284165 0.9990728
#> 824:     151  0.09440699           0        5 -2.0874756 0.036845165 0.9227890
#> 825:     151  0.09440699           0        5 -0.3438946 0.730925532 0.9990728
#> 826:     151  0.09440699           0        5  0.5096071 0.610326718 0.9990728
#>      p.res.onesided fdr.res.onesided rjmcmc rjmcmc_pos rjmcmc_pos_present
#>   1:      0.9643044        0.9993435 0.0007      3e-04          0.4285714
#>   2:      0.2611525        0.9899011 0.0011      6e-04          0.5454545
#>   3:      0.3498673        0.9899011 0.0016      8e-04          0.5000000
#>   4:      0.3532949        0.9899011 0.0014      5e-04          0.3571429
#>   5:      0.9974270        0.9994254 0.0039      0e+00          0.0000000
#>  ---                                                                     
#> 822:      0.3345352        0.9899011 0.0010      5e-04          0.5000000
#> 823:      0.3316421        0.9899011 0.0015      8e-04          0.5333333
#> 824:      0.9815774        0.9993435 0.0011      0e+00          0.0000000
#> 825:      0.6345372        0.9956451 0.0009      2e-04          0.2222222
#> 826:      0.3051634        0.9899011 0.0011      6e-04          0.5454545
#>      rjmcmc_neg rjmcmc_neg_present     beta_mean beta_mean_present beta
#>   1:     0.0004          0.5714286 -7.403422e-05        0.08576166    0
#>   2:     0.0005          0.4545455  9.151022e-05        0.17097391    0
#>   3:     0.0008          0.5000000 -9.539717e-06        0.07722270    0
#>   4:     0.0009          0.6428571 -4.710002e-04        0.18928844    0
#>   5:     0.0039          1.0000000 -5.219369e-03               NaN    0
#>  ---                                                                   
#> 822:     0.0005          0.5000000  5.150655e-05        0.21048863    0
#> 823:     0.0007          0.4666667  2.419349e-04        0.44504593    0
#> 824:     0.0011          1.0000000 -4.827782e-04               NaN    0
#> 825:     0.0007          0.7777778  7.155997e-05        0.67835811    0
#> 826:     0.0005          0.4545455 -2.196721e-05        0.17063607    0
#>      beta_present     predicted predicted_present  pos
#>   1:   0.02102332 -7.403422e-05               NaN   P1
#>   2:   0.08056142  9.151022e-05               NaN   P2
#>   3:   0.05801801 -9.539717e-06               NaN   P3
#>   4:   0.14023309 -4.710002e-04               NaN   P4
#>   5:           NA -5.219369e-03               NaN   P5
#>  ---                                                  
#> 822:   0.17298259  5.150655e-05               NaN P822
#> 823:   0.36202535  2.419349e-04               NaN P823
#> 824:           NA -4.827782e-04               NaN P824
#> 825:   0.67835811  7.155997e-05               NaN P825
#> 826:   0.09633610 -2.196721e-05               NaN P826
#> 

# }