Wrapper for model_bin() that interacts with the filesystem

Parametrizes a negative binomial null model for the readcounts in a given distance bin. Covariates are the distance between an interaction's bait and prey fragments, their lengths, and transchromosomal bait activity. Randomized quantile residuals computed from this model are taken as noise-adjusted readcounts and stored on disk.

model_bin_fs(
  bins_dir,
  bin_index,
  output_dir,
  subsample_size = NA,
  gamlss_cycles = 200,
  gamlss_crit = 0.1,
  formula_add = NA,
  formula_replace = NA
)

Arguments

bins_dir
Directory containing putative interactions that are binned by distance.

bin_index
Index of the bin to be processed.

output_dir
Directory where the parametrized model and adjusted readcounts will be stored. Will be created if it does not exist.

subsample_size
Number of interactions based on which the null-model is parametrized. By default, all are used.

gamlss_cycles
GAMLSS maximum number of cycles for convergence (see gamlss::gamlss.control).

gamlss_crit
GAMLSS convergence criterion (see gamlss::gamlss.control).

formula_add
Additional covariates for the default model formula. To add covariates A and B when modelling counts, supply "A + B".

formula_replace
Replaces the default model formula entirely. Suppled in the form "N ~ A + B". Takes precedence over formula_add.

Value

List containing the path to the output directory, the null model, and the adjusted read counts.

Examples

base = system.file("extdata",package="peaky")
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

bin_interactions_fs(interactions_file, fragments_file, output_dir=bins_dir)

#> 31-08-2020 17:10:11
#> Reading interactions from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/counts.tsv
#> 31-08-2020 17:10:11
#> Reading fragment information from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fragments.bed
#> 31-08-2020 17:10:11
#> Calculating fragment characteristics...
#> 31-08-2020 17:10:11
#> Adding fragment characteristics for baits...
#> 31-08-2020 17:10:11
#> Adding fragment characteristics for preys...
#> 31-08-2020 17:10:11
#> Calculating interaction distances...
#> 31-08-2020 17:10:11
#> Calculating total trans-chromosomal read counts for each bait...
#> 31-08-2020 17:10:11
#> Modelling those as a function of bait chromosome...
#> GAMLSS-RS iteration 1: Global Deviance = 363.9654 
#> GAMLSS-RS iteration 2: Global Deviance = 363.9654 
#> 31-08-2020 17:10:11
#> Adding trans-chromosomal interactivity covariate for preys that were also baited (0 for preys not baited)...
#> 31-08-2020 17:10:11
#> Excluding 2 interactions that are too proximal (distance < 2500 bp)...
#> 31-08-2020 17:10:11
#> Excluding 0 interactions that are too distal (distance > Inf bp)...
#> 31-08-2020 17:10:11
#> Assigning 5 distance bins...
#> 31-08-2020 17:10:11
#> Done.
#> 31-08-2020 17:10:11
#> Saving binned interactions:
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> 31-08-2020 17:10:12
#> Saving bin details to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bins.txt

#> $output_dir
#> [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins"
#> 
#> $interactions
#>        baitID preyID N b.chr     b.mid b.length p.chr     p.mid p.length
#>     1:  53559  52096 1     1 212337456     8519     1 207345336     1932
#>     2:  53559  52107 1     1 212337456     8519     1 207380122      891
#>     3:  53559  52112 1     1 212337456     8519     1 207394962     6067
#>     4:  53559  52113 1     1 212337456     8519     1 207398127      260
#>     5:  53559  52121 1     1 212337456     8519     1 207434420     1859
#>    ---                                                                  
#> 78050: 661722 830195 1     6 170834793     4398     X 155188600      619
#> 78051: 661722 831633 1     6 170834793     4398     Y   7238968     2825
#> 78052: 661722 832798 1     6 170834793     4398     Y  14832472     5436
#> 78053: 661722 833895 1     6 170834793     4398     Y  18220566     2150
#> 78054: 661722 835350 1     6 170834793     4398     Y  22944058     3103
#>            dist b.trans b.trans_res p.trans_res dist.bin
#>     1: -4992120     151  0.09818457           0        5
#>     2: -4957334     151  0.09818457           0        5
#>     3: -4942494     151  0.09818457           0        5
#>     4: -4939329     151  0.09818457           0        5
#>     5: -4903036     151  0.09818457           0        5
#>    ---                                                  
#> 78050:       NA    1719  2.21465890           0     <NA>
#> 78051:       NA    1719  2.21465890           0     <NA>
#> 78052:       NA    1719  2.21465890           0     <NA>
#> 78053:       NA    1719  2.21465890           0     <NA>
#> 78054:       NA    1719  2.21465890           0     <NA>
#> 
#> $bins
#>    dist.bin dist.abs.min dist.abs.max interactions
#> 1:        1         2907       733944         9064
#> 2:        2       734021      1590586         9063
#> 3:        3      1590733      2558841         9063
#> 4:        4      2559052      3664236         9063
#> 5:        5      3664323      4999963         9064
#> 6:     <NA>           NA           NA        32737
#> 

fits_dir = paste0(base,"/fits")

for(bin_index in 1:8){
  model_bin_fs(bins_dir,bin_index,output_dir=fits_dir)
}

#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:10:12
#> No subsampling size provided, using all 9064 interactions for the null model regression...
#> 31-08-2020 17:10:12
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 70224.43 
#> GAMLSS-RS iteration 2: Global Deviance = 69014.53 
#> GAMLSS-RS iteration 3: Global Deviance = 68271.49 
#> GAMLSS-RS iteration 4: Global Deviance = 67988.1 
#> GAMLSS-RS iteration 5: Global Deviance = 67951.65 
#> GAMLSS-RS iteration 6: Global Deviance = 67949.6 
#> GAMLSS-RS iteration 7: Global Deviance = 67949.51 
#> 31-08-2020 17:10:15
#> Converged: YES
#> Iterations: 7
#> 
#> Coefficients:
#> (Intercept)	16.2891915816431
#>  log(abs(dist))	-1.07232040509887
#>  b.trans_res	0.189027735946146
#>  p.trans_res	0.259438430072281
#>  sqrt(b.length)	-0.00374509117431207
#>  sqrt(p.length)	0.00044587597459571
#> 31-08-2020 17:10:15
#> Obtaining normalized randomized quantile residuals for the full dataset...

#> Warning: There is a discrepancy  between the original and the re-fit 
#>  used to achieve 'safe' predictions 
#>  

#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_1.rds...
#> 31-08-2020 17:10:16
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_1.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:10:16
#> No subsampling size provided, using all 9063 interactions for the null model regression...
#> 31-08-2020 17:10:16
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 38496.88 
#> GAMLSS-RS iteration 2: Global Deviance = 38488.12 
#> GAMLSS-RS iteration 3: Global Deviance = 38483.31 
#> GAMLSS-RS iteration 4: Global Deviance = 38480.42 
#> GAMLSS-RS iteration 5: Global Deviance = 38478.57 
#> GAMLSS-RS iteration 6: Global Deviance = 38477.33 
#> GAMLSS-RS iteration 7: Global Deviance = 38476.48 
#> GAMLSS-RS iteration 8: Global Deviance = 38475.87 
#> GAMLSS-RS iteration 9: Global Deviance = 38475.44 
#> GAMLSS-RS iteration 10: Global Deviance = 38475.12 
#> GAMLSS-RS iteration 11: Global Deviance = 38474.88 
#> GAMLSS-RS iteration 12: Global Deviance = 38474.71 
#> GAMLSS-RS iteration 13: Global Deviance = 38474.57 
#> GAMLSS-RS iteration 14: Global Deviance = 38474.47 
#> GAMLSS-RS iteration 15: Global Deviance = 38474.39 
#> 31-08-2020 17:10:21
#> Converged: YES
#> Iterations: 15
#> 
#> Coefficients:
#> (Intercept)	31.2628874318663
#>  log(abs(dist))	-2.18120665282174
#>  b.trans_res	0.420595782469913
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00405883919113198
#>  sqrt(p.length)	-0.000733435309579321
#> 31-08-2020 17:10:21
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_2.rds...
#> 31-08-2020 17:10:21
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_2.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:10:21
#> No subsampling size provided, using all 9063 interactions for the null model regression...
#> 31-08-2020 17:10:21
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 26356.17 
#> GAMLSS-RS iteration 2: Global Deviance = 26196.25 
#> GAMLSS-RS iteration 3: Global Deviance = 26148.22 
#> GAMLSS-RS iteration 4: Global Deviance = 26127.33 
#> GAMLSS-RS iteration 5: Global Deviance = 26116.51 
#> GAMLSS-RS iteration 6: Global Deviance = 26110.3 
#> GAMLSS-RS iteration 7: Global Deviance = 26106.47 
#> GAMLSS-RS iteration 8: Global Deviance = 26103.99 
#> GAMLSS-RS iteration 9: Global Deviance = 26102.32 
#> GAMLSS-RS iteration 10: Global Deviance = 26101.18 
#> GAMLSS-RS iteration 11: Global Deviance = 26100.37 
#> GAMLSS-RS iteration 12: Global Deviance = 26099.78 
#> GAMLSS-RS iteration 13: Global Deviance = 26099.35 
#> GAMLSS-RS iteration 14: Global Deviance = 26099.04 
#> GAMLSS-RS iteration 15: Global Deviance = 26098.81 
#> GAMLSS-RS iteration 16: Global Deviance = 26098.63 
#> GAMLSS-RS iteration 17: Global Deviance = 26098.5 
#> GAMLSS-RS iteration 18: Global Deviance = 26098.4 
#> 31-08-2020 17:10:28
#> Converged: YES
#> Iterations: 18
#> 
#> Coefficients:
#> (Intercept)	34.0696701615016
#>  log(abs(dist))	-2.31223959055367
#>  b.trans_res	0.320599691258825
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00774971586237189
#>  sqrt(p.length)	-0.00225016633634792
#> 31-08-2020 17:10:28
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_3.rds...
#> 31-08-2020 17:10:28
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_3.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:10:28
#> No subsampling size provided, using all 9063 interactions for the null model regression...
#> 31-08-2020 17:10:28
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 19943.82 
#> GAMLSS-RS iteration 2: Global Deviance = 19820.68 
#> GAMLSS-RS iteration 3: Global Deviance = 19780.78 
#> GAMLSS-RS iteration 4: Global Deviance = 19762.73 
#> GAMLSS-RS iteration 5: Global Deviance = 19753.17 
#> GAMLSS-RS iteration 6: Global Deviance = 19747.59 
#> GAMLSS-RS iteration 7: Global Deviance = 19744.12 
#> GAMLSS-RS iteration 8: Global Deviance = 19741.88 
#> GAMLSS-RS iteration 9: Global Deviance = 19740.38 
#> GAMLSS-RS iteration 10: Global Deviance = 19739.34 
#> GAMLSS-RS iteration 11: Global Deviance = 19738.61 
#> GAMLSS-RS iteration 12: Global Deviance = 19738.1 
#> GAMLSS-RS iteration 13: Global Deviance = 19737.73 
#> GAMLSS-RS iteration 14: Global Deviance = 19737.45 
#> GAMLSS-RS iteration 15: Global Deviance = 19737.25 
#> GAMLSS-RS iteration 16: Global Deviance = 19737.1 
#> GAMLSS-RS iteration 17: Global Deviance = 19736.99 
#> GAMLSS-RS iteration 18: Global Deviance = 19736.91 
#> 31-08-2020 17:10:35
#> Converged: YES
#> Iterations: 18
#> 
#> Coefficients:
#> (Intercept)	18.7562310577159
#>  log(abs(dist))	-1.25942194451802
#>  b.trans_res	0.397854327518701
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00627266524613578
#>  sqrt(p.length)	-0.00186343323434686
#> 31-08-2020 17:10:35
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_4.rds...
#> 31-08-2020 17:10:35
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_4.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:10:35
#> No subsampling size provided, using all 9064 interactions for the null model regression...
#> 31-08-2020 17:10:35
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 13972.75 
#> GAMLSS-RS iteration 2: Global Deviance = 13899.99 
#> GAMLSS-RS iteration 3: Global Deviance = 13863.26 
#> GAMLSS-RS iteration 4: Global Deviance = 13841 
#> GAMLSS-RS iteration 5: Global Deviance = 13826.05 
#> GAMLSS-RS iteration 6: Global Deviance = 13815.29 
#> GAMLSS-RS iteration 7: Global Deviance = 13807.2 
#> GAMLSS-RS iteration 8: Global Deviance = 13800.88 
#> GAMLSS-RS iteration 9: Global Deviance = 13795.81 
#> GAMLSS-RS iteration 10: Global Deviance = 13791.65 
#> GAMLSS-RS iteration 11: Global Deviance = 13788.2 
#> GAMLSS-RS iteration 12: Global Deviance = 13785.28 
#> GAMLSS-RS iteration 13: Global Deviance = 13782.79 
#> GAMLSS-RS iteration 14: Global Deviance = 13780.63 
#> GAMLSS-RS iteration 15: Global Deviance = 13778.71 
#> GAMLSS-RS iteration 16: Global Deviance = 13777.02 
#> GAMLSS-RS iteration 17: Global Deviance = 13775.51 
#> GAMLSS-RS iteration 18: Global Deviance = 13774.16 
#> GAMLSS-RS iteration 19: Global Deviance = 13772.99 
#> GAMLSS-RS iteration 20: Global Deviance = 13771.94 
#> GAMLSS-RS iteration 21: Global Deviance = 13771 
#> GAMLSS-RS iteration 22: Global Deviance = 13770.1 
#> GAMLSS-RS iteration 23: Global Deviance = 13769.27 
#> GAMLSS-RS iteration 24: Global Deviance = 13768.51 
#> GAMLSS-RS iteration 25: Global Deviance = 13767.85 
#> GAMLSS-RS iteration 26: Global Deviance = 13767.25 
#> GAMLSS-RS iteration 27: Global Deviance = 13766.71 
#> GAMLSS-RS iteration 28: Global Deviance = 13766.14 
#> GAMLSS-RS iteration 29: Global Deviance = 13765.67 
#> GAMLSS-RS iteration 30: Global Deviance = 13765.23 
#> GAMLSS-RS iteration 31: Global Deviance = 13764.76 
#> GAMLSS-RS iteration 32: Global Deviance = 13764.39 
#> GAMLSS-RS iteration 33: Global Deviance = 13763.97 
#> GAMLSS-RS iteration 34: Global Deviance = 13763.65 
#> GAMLSS-RS iteration 35: Global Deviance = 13763.27 
#> GAMLSS-RS iteration 36: Global Deviance = 13763 
#> GAMLSS-RS iteration 37: Global Deviance = 13762.75 
#> GAMLSS-RS iteration 38: Global Deviance = 13762.53 
#> GAMLSS-RS iteration 39: Global Deviance = 13762.33 
#> GAMLSS-RS iteration 40: Global Deviance = 13762.16 
#> GAMLSS-RS iteration 41: Global Deviance = 13762.02 
#> GAMLSS-RS iteration 42: Global Deviance = 13762 
#> 31-08-2020 17:10:57
#> Converged: YES
#> Iterations: 42
#> 
#> Coefficients:
#> (Intercept)	84.0498382331214
#>  log(abs(dist))	-5.58398803022423
#>  b.trans_res	0.219676258565709
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00102396167462816
#>  sqrt(p.length)	-0.00343055074221772
#> 31-08-2020 17:10:57
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_5.rds...
#> 31-08-2020 17:10:57
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_5.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_6.rds

#> Warning: cannot open compressed file '/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_6.rds', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection