Obtains peak locations and heights based on adjusted readcounts

Fits additive models of peaks of varying strengths in various locations to the adjusted readcounts via RJMCMC, and stores these models on disk.

peaky(
  bait,
  omega_power,
  iterations = 1e+06,
  min_interactions = 20,
  log_file = NA
)

Arguments

bait
Data table containing the putative interactions of a bait, having the columns 'baitID', 'dist', and 'residual'. These report the bait ID, its distance to putative preys, and the adjusted readcounts for its interactions with them, respectively.

omega_power
Expected decay of adjusted read counts around a truly interacting prey. See details.

iterations
Number of models to parametrize. Greated numbers should lead to increased reproducibility.

log_file
Path to a log file.

Value

The output directory.

Details

The steepness of the function to be fitted to putative peaks is determined by \(\omega\) according to \(\beta\) exp(-\(|\omega * d|\)), where \(\beta\) represents peak height and \(d\) the distance from the center of the peak in bp.

Examples

base = system.file("extdata",package="peaky")
interactions_file = paste0(base,"/counts.tsv")
fragments_file = paste0(base,"/fragments.bed")
interactions = data.table::fread(interactions_file)
fragments = data.table::fread(fragments_file)

BI = bin_interactions(interactions, fragments, bins=5)

#> 31-08-2020 17:10:58
#> Calculating fragment characteristics...
#> 31-08-2020 17:10:58
#> Adding fragment characteristics for baits...
#> 31-08-2020 17:10:58
#> Adding fragment characteristics for preys...
#> 31-08-2020 17:10:58
#> Calculating interaction distances...
#> 31-08-2020 17:10:58
#> Calculating total trans-chromosomal read counts for each bait...
#> 31-08-2020 17:10:58
#> Modelling those as a function of bait chromosome...
#> GAMLSS-RS iteration 1: Global Deviance = 363.9654 
#> GAMLSS-RS iteration 2: Global Deviance = 363.9654 
#> 31-08-2020 17:10:58
#> Adding trans-chromosomal interactivity covariate for preys that were also baited (0 for preys not baited)...
#> 31-08-2020 17:10:58
#> Excluding 2 interactions that are too proximal (distance < 2500 bp)...
#> 31-08-2020 17:10:58
#> Excluding 0 interactions that are too distal (distance > Inf bp)...
#> 31-08-2020 17:10:58
#> Assigning 5 distance bins...
#> 31-08-2020 17:10:58
#> Done.

models = by(BI$interactions, BI$interactions$dist.bin, model_bin, subsample_size=1000)

#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:10:58
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 17:10:58
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 7879.118 
#> GAMLSS-RS iteration 2: Global Deviance = 7738.21 
#> GAMLSS-RS iteration 3: Global Deviance = 7687.054 
#> GAMLSS-RS iteration 4: Global Deviance = 7679.895 
#> GAMLSS-RS iteration 5: Global Deviance = 7679.48 
#> GAMLSS-RS iteration 6: Global Deviance = 7679.46 
#> 31-08-2020 17:10:58
#> Converged: YES
#> Iterations: 6
#> 
#> Coefficients:
#> (Intercept)	16.3088460898645
#>  log(abs(dist))	-1.06100198073832
#>  b.trans_res	0.202258431340304
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00546629450025025
#>  sqrt(p.length)	0.000628560430519802
#> 31-08-2020 17:10:58
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:10:58
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:10:58
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 4161.726 
#> GAMLSS-RS iteration 2: Global Deviance = 4161.654 
#> 31-08-2020 17:10:58
#> Converged: YES
#> Iterations: 2
#> 
#> Coefficients:
#> (Intercept)	30.4039774679331
#>  log(abs(dist))	-2.12716561569803
#>  b.trans_res	0.546223086800347
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.0032117909226864
#>  sqrt(p.length)	-0.000209452299999696
#> 31-08-2020 17:10:58
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:10:58
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:10:58
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 3010.195 
#> GAMLSS-RS iteration 2: Global Deviance = 2995.445 
#> GAMLSS-RS iteration 3: Global Deviance = 2990.814 
#> GAMLSS-RS iteration 4: Global Deviance = 2988.766 
#> GAMLSS-RS iteration 5: Global Deviance = 2987.693 
#> GAMLSS-RS iteration 6: Global Deviance = 2987.074 
#> GAMLSS-RS iteration 7: Global Deviance = 2986.692 
#> GAMLSS-RS iteration 8: Global Deviance = 2986.446 
#> GAMLSS-RS iteration 9: Global Deviance = 2986.278 
#> GAMLSS-RS iteration 10: Global Deviance = 2986.164 
#> GAMLSS-RS iteration 11: Global Deviance = 2986.082 
#> 31-08-2020 17:10:59
#> Converged: YES
#> Iterations: 11
#> 
#> Coefficients:
#> (Intercept)	35.1122758962043
#>  log(abs(dist))	-2.3717025282306
#>  b.trans_res	0.288189915115696
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00898830876781525
#>  sqrt(p.length)	-0.00238277553904216
#> 31-08-2020 17:10:59
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:10:59
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:10:59
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 2165.388 
#> GAMLSS-RS iteration 2: Global Deviance = 2156.599 
#> GAMLSS-RS iteration 3: Global Deviance = 2153.597 
#> GAMLSS-RS iteration 4: Global Deviance = 2152.23 
#> GAMLSS-RS iteration 5: Global Deviance = 2151.517 
#> GAMLSS-RS iteration 6: Global Deviance = 2151.108 
#> GAMLSS-RS iteration 7: Global Deviance = 2150.86 
#> GAMLSS-RS iteration 8: Global Deviance = 2150.703 
#> GAMLSS-RS iteration 9: Global Deviance = 2150.603 
#> GAMLSS-RS iteration 10: Global Deviance = 2150.536 
#> 31-08-2020 17:10:59
#> Converged: YES
#> Iterations: 10
#> 
#> Coefficients:
#> (Intercept)	31.4254012242767
#>  log(abs(dist))	-2.10807474035435
#>  b.trans_res	0.389872199461601
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00538820174854753
#>  sqrt(p.length)	-0.000158430710782132
#> 31-08-2020 17:10:59
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:10:59
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 17:10:59
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 1483.035 
#> GAMLSS-RS iteration 2: Global Deviance = 1476.517 
#> GAMLSS-RS iteration 3: Global Deviance = 1473.174 
#> GAMLSS-RS iteration 4: Global Deviance = 1471.14 
#> GAMLSS-RS iteration 5: Global Deviance = 1469.786 
#> GAMLSS-RS iteration 6: Global Deviance = 1468.821 
#> GAMLSS-RS iteration 7: Global Deviance = 1468.096 
#> GAMLSS-RS iteration 8: Global Deviance = 1467.533 
#> GAMLSS-RS iteration 9: Global Deviance = 1467.084 
#> GAMLSS-RS iteration 10: Global Deviance = 1466.725 
#> GAMLSS-RS iteration 11: Global Deviance = 1466.427 
#> GAMLSS-RS iteration 12: Global Deviance = 1466.175 
#> GAMLSS-RS iteration 13: Global Deviance = 1465.961 
#> GAMLSS-RS iteration 14: Global Deviance = 1465.776 
#> GAMLSS-RS iteration 15: Global Deviance = 1465.615 
#> GAMLSS-RS iteration 16: Global Deviance = 1465.475 
#> GAMLSS-RS iteration 17: Global Deviance = 1465.35 
#> GAMLSS-RS iteration 18: Global Deviance = 1465.24 
#> GAMLSS-RS iteration 19: Global Deviance = 1465.144 
#> 31-08-2020 17:11:00
#> Converged: YES
#> Iterations: 19
#> 
#> Coefficients:
#> (Intercept)	61.5478651008403
#>  log(abs(dist))	-4.09517970613365
#>  b.trans_res	0.243458898319126
#>  p.trans_res	NA
#>  sqrt(b.length)	5.45695879948388e-05
#>  sqrt(p.length)	-0.00495699602804624
#> 31-08-2020 17:11:00
#> Obtaining normalized randomized quantile residuals for the full dataset...

residuals = lapply(models, "[[", "residuals")
bins = split(BI$interactions, BI$interactions$dist.bin)

BTS = split_baits(bins, residuals)

#> 31-08-2020 17:11:00
#> Calculating p-values and performing bin-wise FDR-adjustments...

peaky(BTS[baitID==618421], omega_power=-4.7, iterations=10e3)

#> 31-08-2020 17:11:00
#> Constructing distance matrix with omega=10^-4.7...
#> 10000 iterations will be run...
#> Setting user specified argument AlphaPriorMu 
#> Setting user specified argument AlphaPriorSd 
#> Setting user specified argument Alpha_Initial_Value 
#> Setting user specified argument GaussianResidual_Initial_Value 
#> Setting user specified argument GaussianResidualPrior_UnifArg1 
#> Setting user specified argument GaussianResidualPrior_UnifArg2 
#> Setting user specified argument GaussianResidualPriorFamily 
#> Setting user specified argument Add_Move_Probability 
#> Setting user specified argument Delete_Move_Probability 
#> Setting user specified argument Swap_Move_Probability 
#> 0 observations deleted due to missingnessWriting data into an input file for BGLiMS...
#> Data written in 0 hrs 0 mins and 42 seconds.
#> Calling BGLiMS Java algorithm with command: 
#> java -jar "/home/cq/R/x86_64-pc-linux-gnu-library/3.6/R2BGLiMS/BGLiMS/BGLiMS.jar" "/tmp/RtmpBhKk4I/file3c394cb285c3_Aug31171100_Arguments.txt" "/tmp/RtmpBhKk4I/file3c3912b1d09f_Aug31171100_Data.txt" "/tmp/RtmpBhKk4I/file3c3953636a73_Aug31171100_Results.txt" 10000 0 1 1000 127277 1 1 0 -1 1 1 1 1707 
#> BGLiMS Java computation finished in 0 hrs 0 mins and 15 seconds.
#> Reading BGLiMS posterior output...
#> Posterior output processed in 0 hrs 0 mins and 6 seconds.
#> Finished.

#> An object of class R2BGLiMS_Results.
#> Data were analysed for 1707 samples under the Gaussian likelihood.
#> Model selection was performed for 1707 covariates.
#> Inference was obtained by running RJMCMC for 10000 iterations.
#> A posterior summary table of all parameters is contained in the slot 'posterior.summary.table'