Wrapper for peaky() that interacts with the filesystem

Fits additive models of peaks of varying strengths in various locations to the adjusted readcounts via RJMCMC, and stores these models on disk.

peaky_fs(
  baitlist,
  index,
  output_dir,
  omega_power,
  iterations = 1e+06,
  min_interactions = 20
)

Arguments

baitlist
Path to the list of baits to be processed.

index
Which bait on the list to process, with 1 corresponding to the first one on the list.

output_dir
Directory where all RJMCMC results will be stored. Will be created if it does not exist.

omega_power
Expected decay of adjusted read counts around a truly interacting prey. See details.

iterations
Number of models to parametrize. Greated numbers should lead to increased reproducibility.

min_interactions
Minimum requirement for the number of prey fragments (and thus counts) associated with a bait, baits with fewer are skipped.

Value

List containing the model output directory and the models themselves.

Details

The steepness of the function to be fitted to putative peaks is determined by \(\omega\) according to \(\beta\) exp(-\(|\omega * d|\)), where \(\beta\) represents peak height and \(d\) the distance from the center of the peak in bp.

Examples

base = system.file("extdata",package="peaky")
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

bin_interactions_fs(interactions_file, fragments_file, output_dir=bins_dir)

#> 31-08-2020 17:12:03
#> Reading interactions from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/counts.tsv
#> 31-08-2020 17:12:03
#> Reading fragment information from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fragments.bed
#> 31-08-2020 17:12:03
#> Calculating fragment characteristics...
#> 31-08-2020 17:12:03
#> Adding fragment characteristics for baits...
#> 31-08-2020 17:12:03
#> Adding fragment characteristics for preys...
#> 31-08-2020 17:12:03
#> Calculating interaction distances...
#> 31-08-2020 17:12:03
#> Calculating total trans-chromosomal read counts for each bait...
#> 31-08-2020 17:12:03
#> Modelling those as a function of bait chromosome...
#> GAMLSS-RS iteration 1: Global Deviance = 363.9654 
#> GAMLSS-RS iteration 2: Global Deviance = 363.9654 
#> 31-08-2020 17:12:03
#> Adding trans-chromosomal interactivity covariate for preys that were also baited (0 for preys not baited)...
#> 31-08-2020 17:12:03
#> Excluding 2 interactions that are too proximal (distance < 2500 bp)...
#> 31-08-2020 17:12:03
#> Excluding 0 interactions that are too distal (distance > Inf bp)...
#> 31-08-2020 17:12:03
#> Assigning 5 distance bins...
#> 31-08-2020 17:12:03
#> Done.
#> 31-08-2020 17:12:03
#> Saving binned interactions:
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> 31-08-2020 17:12:04
#> Saving bin details to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bins.txt

#> $output_dir
#> [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins"
#> 
#> $interactions
#>        baitID preyID N b.chr     b.mid b.length p.chr     p.mid p.length
#>     1:  53559  52096 1     1 212337456     8519     1 207345336     1932
#>     2:  53559  52107 1     1 212337456     8519     1 207380122      891
#>     3:  53559  52112 1     1 212337456     8519     1 207394962     6067
#>     4:  53559  52113 1     1 212337456     8519     1 207398127      260
#>     5:  53559  52121 1     1 212337456     8519     1 207434420     1859
#>    ---                                                                  
#> 78050: 661722 830195 1     6 170834793     4398     X 155188600      619
#> 78051: 661722 831633 1     6 170834793     4398     Y   7238968     2825
#> 78052: 661722 832798 1     6 170834793     4398     Y  14832472     5436
#> 78053: 661722 833895 1     6 170834793     4398     Y  18220566     2150
#> 78054: 661722 835350 1     6 170834793     4398     Y  22944058     3103
#>            dist b.trans b.trans_res p.trans_res dist.bin
#>     1: -4992120     151  0.09989108           0        5
#>     2: -4957334     151  0.09989108           0        5
#>     3: -4942494     151  0.09989108           0        5
#>     4: -4939329     151  0.09989108           0        5
#>     5: -4903036     151  0.09989108           0        5
#>    ---                                                  
#> 78050:       NA    1719  2.21350086           0     <NA>
#> 78051:       NA    1719  2.21350086           0     <NA>
#> 78052:       NA    1719  2.21350086           0     <NA>
#> 78053:       NA    1719  2.21350086           0     <NA>
#> 78054:       NA    1719  2.21350086           0     <NA>
#> 
#> $bins
#>    dist.bin dist.abs.min dist.abs.max interactions
#> 1:        1         2907       733944         9064
#> 2:        2       734021      1590586         9063
#> 3:        3      1590733      2558841         9063
#> 4:        4      2559052      3664236         9063
#> 5:        5      3664323      4999963         9064
#> 6:     <NA>           NA           NA        32737
#> 

fits_dir = paste0(base,"/fits")

for(bin_index in 1:5){
  model_bin_fs(bins_dir,bin_index,output_dir=fits_dir,subsample_size=1000)
}

#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:12:04
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 17:12:04
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 7815.597 
#> GAMLSS-RS iteration 2: Global Deviance = 7677.173 
#> GAMLSS-RS iteration 3: Global Deviance = 7581.969 
#> GAMLSS-RS iteration 4: Global Deviance = 7546.592 
#> GAMLSS-RS iteration 5: Global Deviance = 7542.676 
#> GAMLSS-RS iteration 6: Global Deviance = 7542.492 
#> GAMLSS-RS iteration 7: Global Deviance = 7542.485 
#> 31-08-2020 17:12:04
#> Converged: YES
#> Iterations: 7
#> 
#> Coefficients:
#> (Intercept)	16.1091995474435
#>  log(abs(dist))	-1.05474420803584
#>  b.trans_res	0.149010759796944
#>  p.trans_res	0.594785405280135
#>  sqrt(b.length)	-0.00257282980834544
#>  sqrt(p.length)	-0.00115972938594376
#> 31-08-2020 17:12:04
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_1.rds...
#> 31-08-2020 17:12:04
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_1.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:12:04
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:12:04
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 4273.107 
#> GAMLSS-RS iteration 2: Global Deviance = 4271.58 
#> GAMLSS-RS iteration 3: Global Deviance = 4270.668 
#> GAMLSS-RS iteration 4: Global Deviance = 4270.05 
#> GAMLSS-RS iteration 5: Global Deviance = 4269.64 
#> GAMLSS-RS iteration 6: Global Deviance = 4269.33 
#> GAMLSS-RS iteration 7: Global Deviance = 4269.118 
#> GAMLSS-RS iteration 8: Global Deviance = 4268.944 
#> GAMLSS-RS iteration 9: Global Deviance = 4268.825 
#> GAMLSS-RS iteration 10: Global Deviance = 4268.736 
#> 31-08-2020 17:12:04
#> Converged: YES
#> Iterations: 10
#> 
#> Coefficients:
#> (Intercept)	35.3737020759213
#>  log(abs(dist))	-2.47271599627921
#>  b.trans_res	0.485858976549372
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00323869543787587
#>  sqrt(p.length)	-0.00236224939619711
#> 31-08-2020 17:12:04
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_2.rds...
#> 31-08-2020 17:12:05
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_2.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:12:05
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:12:05
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 2800.635 
#> GAMLSS-RS iteration 2: Global Deviance = 2782.657 
#> GAMLSS-RS iteration 3: Global Deviance = 2777.573 
#> GAMLSS-RS iteration 4: Global Deviance = 2775.439 
#> GAMLSS-RS iteration 5: Global Deviance = 2774.361 
#> GAMLSS-RS iteration 6: Global Deviance = 2773.764 
#> GAMLSS-RS iteration 7: Global Deviance = 2773.403 
#> GAMLSS-RS iteration 8: Global Deviance = 2773.179 
#> GAMLSS-RS iteration 9: Global Deviance = 2773.031 
#> GAMLSS-RS iteration 10: Global Deviance = 2772.93 
#> GAMLSS-RS iteration 11: Global Deviance = 2772.861 
#> 31-08-2020 17:12:05
#> Converged: YES
#> Iterations: 11
#> 
#> Coefficients:
#> (Intercept)	38.2952517590615
#>  log(abs(dist))	-2.58792972659309
#>  b.trans_res	0.323637061357012
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00925516893746584
#>  sqrt(p.length)	-0.00398507785471592
#> 31-08-2020 17:12:05
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_3.rds...
#> 31-08-2020 17:12:05
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_3.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:12:05
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:12:05
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 2150.894 
#> GAMLSS-RS iteration 2: Global Deviance = 2139.137 
#> GAMLSS-RS iteration 3: Global Deviance = 2135.224 
#> GAMLSS-RS iteration 4: Global Deviance = 2133.457 
#> GAMLSS-RS iteration 5: Global Deviance = 2132.527 
#> GAMLSS-RS iteration 6: Global Deviance = 2131.995 
#> GAMLSS-RS iteration 7: Global Deviance = 2131.668 
#> GAMLSS-RS iteration 8: Global Deviance = 2131.457 
#> GAMLSS-RS iteration 9: Global Deviance = 2131.319 
#> GAMLSS-RS iteration 10: Global Deviance = 2131.228 
#> 31-08-2020 17:12:06
#> Converged: YES
#> Iterations: 10
#> 
#> Coefficients:
#> (Intercept)	15.1690825813994
#>  log(abs(dist))	-1.01842062174304
#>  b.trans_res	0.382423316510056
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00605572463252533
#>  sqrt(p.length)	-0.00158026482449677
#> 31-08-2020 17:12:06
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_4.rds...
#> 31-08-2020 17:12:06
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_4.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:12:06
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 17:12:06
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 1515.233 
#> GAMLSS-RS iteration 2: Global Deviance = 1506.049 
#> GAMLSS-RS iteration 3: Global Deviance = 1501.387 
#> GAMLSS-RS iteration 4: Global Deviance = 1498.519 
#> GAMLSS-RS iteration 5: Global Deviance = 1496.578 
#> GAMLSS-RS iteration 6: Global Deviance = 1495.161 
#> GAMLSS-RS iteration 7: Global Deviance = 1494.078 
#> GAMLSS-RS iteration 8: Global Deviance = 1493.227 
#> GAMLSS-RS iteration 9: Global Deviance = 1492.532 
#> GAMLSS-RS iteration 10: Global Deviance = 1491.95 
#> GAMLSS-RS iteration 11: Global Deviance = 1491.466 
#> GAMLSS-RS iteration 12: Global Deviance = 1491.051 
#> GAMLSS-RS iteration 13: Global Deviance = 1490.693 
#> GAMLSS-RS iteration 14: Global Deviance = 1490.38 
#> GAMLSS-RS iteration 15: Global Deviance = 1490.088 
#> GAMLSS-RS iteration 16: Global Deviance = 1489.824 
#> GAMLSS-RS iteration 17: Global Deviance = 1489.584 
#> GAMLSS-RS iteration 18: Global Deviance = 1489.386 
#> GAMLSS-RS iteration 19: Global Deviance = 1489.212 
#> GAMLSS-RS iteration 20: Global Deviance = 1489.056 
#> GAMLSS-RS iteration 21: Global Deviance = 1488.884 
#> GAMLSS-RS iteration 22: Global Deviance = 1488.722 
#> GAMLSS-RS iteration 23: Global Deviance = 1488.608 
#> GAMLSS-RS iteration 24: Global Deviance = 1488.461 
#> GAMLSS-RS iteration 25: Global Deviance = 1488.372 
#> 31-08-2020 17:12:07
#> Converged: YES
#> Iterations: 25
#> 
#> Coefficients:
#> (Intercept)	75.3929016894377
#>  log(abs(dist))	-5.00700707545572
#>  b.trans_res	0.192928456741686
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.0034919058260532
#>  sqrt(p.length)	-0.00333874228312689
#> 31-08-2020 17:12:07
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_5.rds...
#> 31-08-2020 17:12:07
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_5.rds...

baits_dir = paste0(base,"/baits")

split_baits_fs(bins_dir,residuals_dir = fits_dir, indices=1:5, output_dir = baits_dir)

#> 31-08-2020 17:12:07
#> Loading bins...
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> 31-08-2020 17:12:07
#> Loading residuals...
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_1.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_2.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_3.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_4.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_5.rds
#> 31-08-2020 17:12:07
#> Calculating p-values and performing bin-wise FDR-adjustments...
#> 31-08-2020 17:12:07
#> Saving baits...
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_53559.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_71638.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_95958.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_96180.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_102201.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119435.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119522.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_144385.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_162008.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_175487.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_203500.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_243669.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_262362.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_268979.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_272961.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_312780.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_328019.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_341315.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_363262.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_366588.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_466090.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_582536.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_618421.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_658681.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_661722.rds
#> 31-08-2020 17:12:07
#> Plotting baits...

#> 31-08-2020 17:12:10
#> Done.

#> $output_dir
#> [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits"
#> 
#> $bait_paths
#>  [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_53559.rds" 
#>  [2] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_71638.rds" 
#>  [3] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_95958.rds" 
#>  [4] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_96180.rds" 
#>  [5] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_102201.rds"
#>  [6] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119435.rds"
#>  [7] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119522.rds"
#>  [8] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_144385.rds"
#>  [9] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_162008.rds"
#> [10] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_175487.rds"
#> [11] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_203500.rds"
#> [12] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_243669.rds"
#> [13] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_262362.rds"
#> [14] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_268979.rds"
#> [15] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_272961.rds"
#> [16] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_312780.rds"
#> [17] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_328019.rds"
#> [18] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_341315.rds"
#> [19] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_363262.rds"
#> [20] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_366588.rds"
#> [21] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_466090.rds"
#> [22] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_582536.rds"
#> [23] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_618421.rds"
#> [24] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_658681.rds"
#> [25] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_661722.rds"
#> 

baitlist = paste0(baits_dir,"/baitlist.txt")
rjmcmc_dir = paste0(base,"/rjmcmc")
omega_power = -4.7
# \donttest{
for(i in 1:3){
 peaky_fs(baitlist,i,output_dir=rjmcmc_dir,omega_power=omega_power)
}

#> 31-08-2020 17:12:10
#> Loaded bait 53559 from path: /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_53559.rds...
#> 31-08-2020 17:12:10
#> Constructing distance matrix with omega=10^-4.7...
#> 1e+06 iterations will be run...
#> Setting user specified argument AlphaPriorMu 
#> Setting user specified argument AlphaPriorSd 
#> Setting user specified argument Alpha_Initial_Value 
#> Setting user specified argument GaussianResidual_Initial_Value 
#> Setting user specified argument GaussianResidualPrior_UnifArg1 
#> Setting user specified argument GaussianResidualPrior_UnifArg2 
#> Setting user specified argument GaussianResidualPriorFamily 
#> Setting user specified argument Add_Move_Probability 
#> Setting user specified argument Delete_Move_Probability 
#> Setting user specified argument Swap_Move_Probability 
#> 0 observations deleted due to missingnessWriting data into an input file for BGLiMS...
#> Data written in 0 hrs 0 mins and 8 seconds.
#> Calling BGLiMS Java algorithm with command: 
#> java -jar "/home/cq/R/x86_64-pc-linux-gnu-library/3.6/R2BGLiMS/BGLiMS/BGLiMS.jar" "/tmp/RtmpBhKk4I/file3c391c94a51c_Aug31171210_Arguments.txt" "/tmp/RtmpBhKk4I/file3c3926173669_Aug31171210_Data.txt" "/tmp/RtmpBhKk4I/file3c398310c21_Aug31171210_Results.txt" 1000000 0 100 100000 954706 1 1 0 -1 1 1 1 826 
#> BGLiMS Java computation finished in 0 hrs 2 mins and 33 seconds.
#> Reading BGLiMS posterior output...
#> Posterior output processed in 0 hrs 0 mins and 3 seconds.
#> Finished.
#> 31-08-2020 17:14:54
#> Saving RJMCMC results to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc/rjmcmc_53559.rds
#> 31-08-2020 17:14:54
#> Done.
#> 31-08-2020 17:14:54
#> Loaded bait 71638 from path: /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_71638.rds...
#> 31-08-2020 17:14:54
#> Constructing distance matrix with omega=10^-4.7...
#> 1e+06 iterations will be run...
#> Setting user specified argument AlphaPriorMu 
#> Setting user specified argument AlphaPriorSd 
#> Setting user specified argument Alpha_Initial_Value 
#> Setting user specified argument GaussianResidual_Initial_Value 
#> Setting user specified argument GaussianResidualPrior_UnifArg1 
#> Setting user specified argument GaussianResidualPrior_UnifArg2 
#> Setting user specified argument GaussianResidualPriorFamily 
#> Setting user specified argument Add_Move_Probability 
#> Setting user specified argument Delete_Move_Probability 
#> Setting user specified argument Swap_Move_Probability 
#> 0 observations deleted due to missingnessWriting data into an input file for BGLiMS...
#> Data written in 0 hrs 1 mins and 11 seconds.
#> Calling BGLiMS Java algorithm with command: 
#> java -jar "/home/cq/R/x86_64-pc-linux-gnu-library/3.6/R2BGLiMS/BGLiMS/BGLiMS.jar" "/tmp/RtmpBhKk4I/file3c3969472ae0_Aug31171454_Arguments.txt" "/tmp/RtmpBhKk4I/file3c3938c90708_Aug31171454_Data.txt" "/tmp/RtmpBhKk4I/file3c395b947694_Aug31171454_Results.txt" 1000000 0 100 100000 500295 1 1 0 -1 1 1 1 2014 
#> BGLiMS Java computation finished in 0 hrs 6 mins and 25 seconds.
#> Reading BGLiMS posterior output...
#> Posterior output processed in 0 hrs 0 mins and 8 seconds.
#> Finished.
#> 31-08-2020 17:22:38
#> Saving RJMCMC results to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc/rjmcmc_71638.rds
#> 31-08-2020 17:22:39
#> Done.
#> 31-08-2020 17:22:39
#> Loaded bait 95958 from path: /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_95958.rds...
#> 31-08-2020 17:22:39
#> Constructing distance matrix with omega=10^-4.7...
#> 1e+06 iterations will be run...
#> Setting user specified argument AlphaPriorMu 
#> Setting user specified argument AlphaPriorSd 
#> Setting user specified argument Alpha_Initial_Value 
#> Setting user specified argument GaussianResidual_Initial_Value 
#> Setting user specified argument GaussianResidualPrior_UnifArg1 
#> Setting user specified argument GaussianResidualPrior_UnifArg2 
#> Setting user specified argument GaussianResidualPriorFamily 
#> Setting user specified argument Add_Move_Probability 
#> Setting user specified argument Delete_Move_Probability 
#> Setting user specified argument Swap_Move_Probability 
#> 0 observations deleted due to missingnessWriting data into an input file for BGLiMS...
#> Data written in 0 hrs 1 mins and 1 seconds.
#> Calling BGLiMS Java algorithm with command: 
#> java -jar "/home/cq/R/x86_64-pc-linux-gnu-library/3.6/R2BGLiMS/BGLiMS/BGLiMS.jar" "/tmp/RtmpBhKk4I/file3c3942bdd9a5_Aug31172240_Arguments.txt" "/tmp/RtmpBhKk4I/file3c393d1e0b05_Aug31172240_Data.txt" "/tmp/RtmpBhKk4I/file3c392c44da3e_Aug31172240_Results.txt" 1000000 0 100 100000 550541 1 1 0 -1 1 1 1 2037 
#> BGLiMS Java computation finished in 0 hrs 5 mins and 22 seconds.
#> Reading BGLiMS posterior output...
#> Posterior output processed in 0 hrs 0 mins and 6 seconds.
#> Finished.
#> 31-08-2020 17:29:09
#> Saving RJMCMC results to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/rjmcmc/rjmcmc_95958.rds
#> 31-08-2020 17:29:10
#> Done.

# }