Regroups putative interactions by bait and calculates p-values under the negative binomial model

Merges bin-wise data, calculates the p-values for putative interactions under the negative binomial model. Regroups putative interactions by bait rather than by distance bin.

split_baits(bins, residuals, log_file = NA)

Arguments

bins
List of data tables containing putative interactions that are binned by distance.

residuals
List containing residuals (adjusted read counts) for each bin (matching the interaction order).

log_file
Path to a log file.

Value

Data table with p-values, residuals and putative interaction attributes that can be sorted or split by bait ID.

Examples

base = system.file("extdata",package="peaky")
interactions_file = paste0(base,"/counts.tsv")
fragments_file = paste0(base,"/fragments.bed")
interactions = data.table::fread(interactions_file)
fragments = data.table::fread(fragments_file)

BI = bin_interactions(interactions, fragments, bins=5)

#> 31-08-2020 17:45:53
#> Calculating fragment characteristics...
#> 31-08-2020 17:45:53
#> Adding fragment characteristics for baits...
#> 31-08-2020 17:45:53
#> Adding fragment characteristics for preys...
#> 31-08-2020 17:45:53
#> Calculating interaction distances...
#> 31-08-2020 17:45:53
#> Calculating total trans-chromosomal read counts for each bait...
#> 31-08-2020 17:45:53
#> Modelling those as a function of bait chromosome...
#> GAMLSS-RS iteration 1: Global Deviance = 363.9654 
#> GAMLSS-RS iteration 2: Global Deviance = 363.9654 
#> 31-08-2020 17:45:53
#> Adding trans-chromosomal interactivity covariate for preys that were also baited (0 for preys not baited)...
#> 31-08-2020 17:45:53
#> Excluding 2 interactions that are too proximal (distance < 2500 bp)...
#> 31-08-2020 17:45:53
#> Excluding 0 interactions that are too distal (distance > Inf bp)...
#> 31-08-2020 17:45:53
#> Assigning 5 distance bins...
#> 31-08-2020 17:45:53
#> Done.

models = by(BI$interactions, BI$interactions$dist.bin, model_bin, subsample_size=1000)

#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:45:53
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 17:45:53
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 7729.572 
#> GAMLSS-RS iteration 2: Global Deviance = 7603.166 
#> GAMLSS-RS iteration 3: Global Deviance = 7524.902 
#> GAMLSS-RS iteration 4: Global Deviance = 7496.745 
#> GAMLSS-RS iteration 5: Global Deviance = 7493.202 
#> GAMLSS-RS iteration 6: Global Deviance = 7492.997 
#> GAMLSS-RS iteration 7: Global Deviance = 7492.989 
#> 31-08-2020 17:45:53
#> Converged: YES
#> Iterations: 7
#> 
#> Coefficients:
#> (Intercept)	16.2479133425831
#>  log(abs(dist))	-1.07780567327871
#>  b.trans_res	0.256673189320444
#>  p.trans_res	0.713045017132021
#>  sqrt(b.length)	-0.00348434717901164
#>  sqrt(p.length)	0.00272871373749522
#> 31-08-2020 17:45:53
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:45:53
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:45:53
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 4396.391 
#> GAMLSS-RS iteration 2: Global Deviance = 4395.574 
#> GAMLSS-RS iteration 3: Global Deviance = 4395.134 
#> GAMLSS-RS iteration 4: Global Deviance = 4394.877 
#> GAMLSS-RS iteration 5: Global Deviance = 4394.699 
#> GAMLSS-RS iteration 6: Global Deviance = 4394.587 
#> GAMLSS-RS iteration 7: Global Deviance = 4394.511 
#> 31-08-2020 17:45:54
#> Converged: YES
#> Iterations: 7
#> 
#> Coefficients:
#> (Intercept)	25.5891570250652
#>  log(abs(dist))	-1.78314349704145
#>  b.trans_res	0.367814283340709
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00480238002129164
#>  sqrt(p.length)	0.00543516656512004
#> 31-08-2020 17:45:54
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:45:54
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:45:54
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 2917.538 
#> GAMLSS-RS iteration 2: Global Deviance = 2900.274 
#> GAMLSS-RS iteration 3: Global Deviance = 2895.745 
#> GAMLSS-RS iteration 4: Global Deviance = 2893.977 
#> GAMLSS-RS iteration 5: Global Deviance = 2893.155 
#> GAMLSS-RS iteration 6: Global Deviance = 2892.726 
#> GAMLSS-RS iteration 7: Global Deviance = 2892.488 
#> GAMLSS-RS iteration 8: Global Deviance = 2892.351 
#> GAMLSS-RS iteration 9: Global Deviance = 2892.268 
#> 31-08-2020 17:45:54
#> Converged: YES
#> Iterations: 9
#> 
#> Coefficients:
#> (Intercept)	22.120981126091
#>  log(abs(dist))	-1.48654672712053
#>  b.trans_res	0.39047058782789
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00437162714361699
#>  sqrt(p.length)	-0.00430773779595917
#> 31-08-2020 17:45:54
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:45:54
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:45:54
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 2259.816 
#> GAMLSS-RS iteration 2: Global Deviance = 2244.552 
#> GAMLSS-RS iteration 3: Global Deviance = 2239.521 
#> GAMLSS-RS iteration 4: Global Deviance = 2237.204 
#> GAMLSS-RS iteration 5: Global Deviance = 2235.94 
#> GAMLSS-RS iteration 6: Global Deviance = 2235.191 
#> GAMLSS-RS iteration 7: Global Deviance = 2234.713 
#> GAMLSS-RS iteration 8: Global Deviance = 2234.396 
#> GAMLSS-RS iteration 9: Global Deviance = 2234.177 
#> GAMLSS-RS iteration 10: Global Deviance = 2234.026 
#> GAMLSS-RS iteration 11: Global Deviance = 2233.915 
#> GAMLSS-RS iteration 12: Global Deviance = 2233.834 
#> 31-08-2020 17:45:54
#> Converged: YES
#> Iterations: 12
#> 
#> Coefficients:
#> (Intercept)	17.6542915305355
#>  log(abs(dist))	-1.18127720882531
#>  b.trans_res	0.321192128192237
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00574450493081633
#>  sqrt(p.length)	-0.00221050699864148
#> 31-08-2020 17:45:54
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:45:54
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 17:45:54
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 1605.029 
#> GAMLSS-RS iteration 2: Global Deviance = 1593.393 
#> GAMLSS-RS iteration 3: Global Deviance = 1587.65 
#> GAMLSS-RS iteration 4: Global Deviance = 1584.179 
#> GAMLSS-RS iteration 5: Global Deviance = 1581.845 
#> GAMLSS-RS iteration 6: Global Deviance = 1580.143 
#> GAMLSS-RS iteration 7: Global Deviance = 1578.838 
#> GAMLSS-RS iteration 8: Global Deviance = 1577.8 
#> GAMLSS-RS iteration 9: Global Deviance = 1576.968 
#> GAMLSS-RS iteration 10: Global Deviance = 1576.289 
#> GAMLSS-RS iteration 11: Global Deviance = 1575.719 
#> GAMLSS-RS iteration 12: Global Deviance = 1575.206 
#> GAMLSS-RS iteration 13: Global Deviance = 1574.756 
#> GAMLSS-RS iteration 14: Global Deviance = 1574.389 
#> GAMLSS-RS iteration 15: Global Deviance = 1574.07 
#> GAMLSS-RS iteration 16: Global Deviance = 1573.751 
#> GAMLSS-RS iteration 17: Global Deviance = 1573.502 
#> GAMLSS-RS iteration 18: Global Deviance = 1573.236 
#> GAMLSS-RS iteration 19: Global Deviance = 1572.992 
#> GAMLSS-RS iteration 20: Global Deviance = 1572.827 
#> GAMLSS-RS iteration 21: Global Deviance = 1572.691 
#> GAMLSS-RS iteration 22: Global Deviance = 1572.581 
#> GAMLSS-RS iteration 23: Global Deviance = 1572.421 
#> GAMLSS-RS iteration 24: Global Deviance = 1572.272 
#> GAMLSS-RS iteration 25: Global Deviance = 1572.144 
#> GAMLSS-RS iteration 26: Global Deviance = 1572.026 
#> GAMLSS-RS iteration 27: Global Deviance = 1571.919 
#> GAMLSS-RS iteration 28: Global Deviance = 1571.821 
#> 31-08-2020 17:45:56
#> Converged: YES
#> Iterations: 28
#> 
#> Coefficients:
#> (Intercept)	100.947091451874
#>  log(abs(dist))	-6.68898523008352
#>  b.trans_res	0.246976742891453
#>  p.trans_res	NA
#>  sqrt(b.length)	0.000130834967813917
#>  sqrt(p.length)	-0.00499682088716204
#> 31-08-2020 17:45:56
#> Obtaining normalized randomized quantile residuals for the full dataset...

#> Warning: There is a discrepancy  between the original and the re-fit 
#>  used to achieve 'safe' predictions 
#>  

residuals = lapply(models, "[[", "residuals")
bins = split(BI$interactions, BI$interactions$dist.bin)

split_baits(bins, residuals)

#> 31-08-2020 17:45:56
#> Calculating p-values and performing bin-wise FDR-adjustments...

#>        baitID preyID N b.chr     b.mid b.length p.chr     p.mid p.length
#>     1:  53559  53370 1     1 212337456     8519     1 211609062      307
#>     2:  53559  53372 1     1 212337456     8519     1 211609866     1285
#>     3:  53559  53374 1     1 212337456     8519     1 211615263     8852
#>     4:  53559  53375 1     1 212337456     8519     1 211622540     5700
#>     5:  53559  53376 1     1 212337456     8519     1 211626450     2118
#>    ---                                                                  
#> 45313: 661722 660774 3     6 170834793     4398     6 167155364     5956
#> 45314: 661722 660775 1     6 170834793     4398     6 167159478     2271
#> 45315: 661722 660776 1     6 170834793     4398     6 167160840      450
#> 45316: 661722 660777 3     6 170834793     4398     6 167162212     2292
#> 45317: 661722 660778 1     6 170834793     4398     6 167167588     8458
#>            dist b.trans b.trans_res p.trans_res dist.bin   residual     p.res
#>     1:  -728394     151  0.09739768           0        1 -1.0326969 0.3017458
#>     2:  -727590     151  0.09739768           0        1 -1.5817786 0.1137001
#>     3:  -722193     151  0.09739768           0        1 -1.3138931 0.1888822
#>     4:  -714916     151  0.09739768           0        1 -1.3071273 0.1911695
#>     5:  -711006     151  0.09739768           0        1 -0.8899715 0.3734812
#>    ---                                                                       
#> 45313: -3679429    1719  2.21367626           0        5  1.3221275 0.1861257
#> 45314: -3675315    1719  2.21367626           0        5 -0.3857964 0.6996474
#> 45315: -3673953    1719  2.21367626           0        5 -0.5197302 0.6032517
#> 45316: -3672581    1719  2.21367626           0        5  0.9121617 0.3616836
#> 45317: -3667205    1719  2.21367626           0        5 -1.5749910 0.1152585
#>          fdr.res p.res.onesided fdr.res.onesided
#>     1: 0.9999400     0.84912711        0.9823390
#>     2: 0.9999400     0.94314994        0.9889763
#>     3: 0.9999400     0.90555891        0.9835533
#>     4: 0.9999400     0.90441526        0.9835533
#>     5: 0.9999400     0.81325941        0.9823390
#>    ---                                          
#> 45313: 0.9955251     0.09306284        0.9998700
#> 45314: 0.9955251     0.65017628        0.9998700
#> 45315: 0.9955251     0.69837417        0.9998700
#> 45316: 0.9955251     0.18084179        0.9998700
#> 45317: 0.9955251     0.94237074        0.9998700