Wrapper for split_baits() that interacts with the filesystem

Calculates p-values under the negative binomial model. Subsequently regroups putative interactions by bait, rather than by distance, to prepare them for parallel RJMCMC processing. Generates a separate file for each bait. Paths are stored in baitlist.txt, which serves as a to-do list for peaky().

split_baits_fs(bins_dir, residuals_dir, indices, output_dir, plots = TRUE)

Arguments

bins_dir
Directory containing putative interactions that are binned by distance.

residuals_dir
Directory where the adjusted read counts from each distance bin are stored.

indices
Indices of distance bins whose baits are processed. These must all have had null models fitted.

output_dir
Directory where all putative interactive will be stored, one file per bait. Will be created if it does not exist.

plots
Whether adjusted readcounts are to be plotted aganst distance and stored for each bait.

Value

The output directory.

Examples

base = system.file("extdata",package="peaky")
interactions_file = paste0(base,"/counts.tsv")
bins_dir = paste0(base,"/bins")
fragments_file = paste0(base,"/fragments.bed")

bin_interactions_fs(interactions_file, fragments_file, output_dir=bins_dir)

#> 31-08-2020 17:45:56
#> Reading interactions from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/counts.tsv
#> 31-08-2020 17:45:56
#> Reading fragment information from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fragments.bed
#> 31-08-2020 17:45:56
#> Calculating fragment characteristics...
#> 31-08-2020 17:45:56
#> Adding fragment characteristics for baits...
#> 31-08-2020 17:45:56
#> Adding fragment characteristics for preys...
#> 31-08-2020 17:45:56
#> Calculating interaction distances...
#> 31-08-2020 17:45:56
#> Calculating total trans-chromosomal read counts for each bait...
#> 31-08-2020 17:45:56
#> Modelling those as a function of bait chromosome...
#> GAMLSS-RS iteration 1: Global Deviance = 363.9654 
#> GAMLSS-RS iteration 2: Global Deviance = 363.9654 
#> 31-08-2020 17:45:56
#> Adding trans-chromosomal interactivity covariate for preys that were also baited (0 for preys not baited)...
#> 31-08-2020 17:45:56
#> Excluding 2 interactions that are too proximal (distance < 2500 bp)...
#> 31-08-2020 17:45:56
#> Excluding 0 interactions that are too distal (distance > Inf bp)...
#> 31-08-2020 17:45:56
#> Assigning 5 distance bins...
#> 31-08-2020 17:45:56
#> Done.
#> 31-08-2020 17:45:56
#> Saving binned interactions:
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> 31-08-2020 17:45:56
#> Saving bin details to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bins.txt

#> $output_dir
#> [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins"
#> 
#> $interactions
#>        baitID preyID N b.chr     b.mid b.length p.chr     p.mid p.length
#>     1:  53559  52096 1     1 212337456     8519     1 207345336     1932
#>     2:  53559  52107 1     1 212337456     8519     1 207380122      891
#>     3:  53559  52112 1     1 212337456     8519     1 207394962     6067
#>     4:  53559  52113 1     1 212337456     8519     1 207398127      260
#>     5:  53559  52121 1     1 212337456     8519     1 207434420     1859
#>    ---                                                                  
#> 78050: 661722 830195 1     6 170834793     4398     X 155188600      619
#> 78051: 661722 831633 1     6 170834793     4398     Y   7238968     2825
#> 78052: 661722 832798 1     6 170834793     4398     Y  14832472     5436
#> 78053: 661722 833895 1     6 170834793     4398     Y  18220566     2150
#> 78054: 661722 835350 1     6 170834793     4398     Y  22944058     3103
#>            dist b.trans b.trans_res p.trans_res dist.bin
#>     1: -4992120     151  0.09415579           0        5
#>     2: -4957334     151  0.09415579           0        5
#>     3: -4942494     151  0.09415579           0        5
#>     4: -4939329     151  0.09415579           0        5
#>     5: -4903036     151  0.09415579           0        5
#>    ---                                                  
#> 78050:       NA    1719  2.21305951           0     <NA>
#> 78051:       NA    1719  2.21305951           0     <NA>
#> 78052:       NA    1719  2.21305951           0     <NA>
#> 78053:       NA    1719  2.21305951           0     <NA>
#> 78054:       NA    1719  2.21305951           0     <NA>
#> 
#> $bins
#>    dist.bin dist.abs.min dist.abs.max interactions
#> 1:        1         2907       733944         9064
#> 2:        2       734021      1590586         9063
#> 3:        3      1590733      2558841         9063
#> 4:        4      2559052      3664236         9063
#> 5:        5      3664323      4999963         9064
#> 6:     <NA>           NA           NA        32737
#> 

fits_dir = paste0(base,"/fits")

for(bin_index in 1:5){
  model_bin_fs(bins_dir,bin_index,output_dir=fits_dir,subsample_size=1000)
}

#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:45:56
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 17:45:56
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 7817.963 
#> GAMLSS-RS iteration 2: Global Deviance = 7671.749 
#> GAMLSS-RS iteration 3: Global Deviance = 7574.145 
#> GAMLSS-RS iteration 4: Global Deviance = 7496.973 
#> GAMLSS-RS iteration 5: Global Deviance = 7465.61 
#> GAMLSS-RS iteration 6: Global Deviance = 7460.643 
#> GAMLSS-RS iteration 7: Global Deviance = 7460.274 
#> GAMLSS-RS iteration 8: Global Deviance = 7460.255 
#> 31-08-2020 17:45:57
#> Converged: YES
#> Iterations: 8
#> 
#> Coefficients:
#> (Intercept)	16.5385014373579
#>  log(abs(dist))	-1.09767977931229
#>  b.trans_res	0.214337600774251
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.0042467636888258
#>  sqrt(p.length)	0.00106936002691906
#> 31-08-2020 17:45:57
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_1.rds...
#> 31-08-2020 17:45:57
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_1.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:45:57
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:45:57
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 4240.508 
#> GAMLSS-RS iteration 2: Global Deviance = 4239.669 
#> GAMLSS-RS iteration 3: Global Deviance = 4239.251 
#> GAMLSS-RS iteration 4: Global Deviance = 4239.02 
#> GAMLSS-RS iteration 5: Global Deviance = 4238.883 
#> GAMLSS-RS iteration 6: Global Deviance = 4238.799 
#> 31-08-2020 17:45:57
#> Converged: YES
#> Iterations: 6
#> 
#> Coefficients:
#> (Intercept)	18.4602692860144
#>  log(abs(dist))	-1.25920693941305
#>  b.trans_res	0.475914727313967
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00491030902249689
#>  sqrt(p.length)	-0.00216378230584008
#> 31-08-2020 17:45:57
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_2.rds...
#> 31-08-2020 17:45:57
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_2.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:45:57
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:45:57
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 2838.83 
#> GAMLSS-RS iteration 2: Global Deviance = 2824.608 
#> GAMLSS-RS iteration 3: Global Deviance = 2821.006 
#> GAMLSS-RS iteration 4: Global Deviance = 2819.653 
#> GAMLSS-RS iteration 5: Global Deviance = 2819.041 
#> GAMLSS-RS iteration 6: Global Deviance = 2818.739 
#> GAMLSS-RS iteration 7: Global Deviance = 2818.578 
#> GAMLSS-RS iteration 8: Global Deviance = 2818.49 
#> 31-08-2020 17:45:57
#> Converged: YES
#> Iterations: 8
#> 
#> Coefficients:
#> (Intercept)	25.2672078610614
#>  log(abs(dist))	-1.71219170142738
#>  b.trans_res	0.356258809942039
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.00467736063108066
#>  sqrt(p.length)	-0.000870934586261226
#> 31-08-2020 17:45:57
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_3.rds...
#> 31-08-2020 17:45:57
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_3.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:45:57
#> Subsampling 1000/9063 interactions for the null model regression...
#> 31-08-2020 17:45:57
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 2219.521 
#> GAMLSS-RS iteration 2: Global Deviance = 2206.709 
#> GAMLSS-RS iteration 3: Global Deviance = 2202.396 
#> GAMLSS-RS iteration 4: Global Deviance = 2200.421 
#> GAMLSS-RS iteration 5: Global Deviance = 2199.363 
#> GAMLSS-RS iteration 6: Global Deviance = 2198.735 
#> GAMLSS-RS iteration 7: Global Deviance = 2198.34 
#> GAMLSS-RS iteration 8: Global Deviance = 2198.086 
#> GAMLSS-RS iteration 9: Global Deviance = 2197.912 
#> GAMLSS-RS iteration 10: Global Deviance = 2197.79 
#> GAMLSS-RS iteration 11: Global Deviance = 2197.702 
#> 31-08-2020 17:45:58
#> Converged: YES
#> Iterations: 11
#> 
#> Coefficients:
#> (Intercept)	22.6678610828902
#>  log(abs(dist))	-1.51806894743518
#>  b.trans_res	0.378018333438826
#>  p.trans_res	NA
#>  sqrt(b.length)	-0.0071601328775684
#>  sqrt(p.length)	-0.000474388960665037
#> 31-08-2020 17:45:58
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_4.rds...
#> 31-08-2020 17:45:58
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_4.rds...
#> Loading bin from /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> A truncated family of distributions from NBI has been generated 
#>  and saved under the names:  
#>  dNBI.0tr pNBI.0tr qNBI.0tr rNBI.0tr NBI.0tr 
#> The type of truncation is left 
#>  and the truncation parameter is 0  
#> 31-08-2020 17:45:58
#> Subsampling 1000/9064 interactions for the null model regression...
#> 31-08-2020 17:45:58
#> Fitting with a maximum of 200 iterations...
#> Using formula:
#> N ~ log(abs(dist)) + b.trans_res  + p.trans_res + sqrt(b.length) + sqrt(p.length)
#> GAMLSS-RS iteration 1: Global Deviance = 1608.663 
#> GAMLSS-RS iteration 2: Global Deviance = 1597.959 
#> GAMLSS-RS iteration 3: Global Deviance = 1592.642 
#> GAMLSS-RS iteration 4: Global Deviance = 1589.437 
#> GAMLSS-RS iteration 5: Global Deviance = 1587.283 
#> GAMLSS-RS iteration 6: Global Deviance = 1585.747 
#> GAMLSS-RS iteration 7: Global Deviance = 1584.586 
#> GAMLSS-RS iteration 8: Global Deviance = 1583.679 
#> GAMLSS-RS iteration 9: Global Deviance = 1582.96 
#> GAMLSS-RS iteration 10: Global Deviance = 1582.375 
#> GAMLSS-RS iteration 11: Global Deviance = 1581.899 
#> GAMLSS-RS iteration 12: Global Deviance = 1581.499 
#> GAMLSS-RS iteration 13: Global Deviance = 1581.161 
#> GAMLSS-RS iteration 14: Global Deviance = 1580.851 
#> GAMLSS-RS iteration 15: Global Deviance = 1580.579 
#> GAMLSS-RS iteration 16: Global Deviance = 1580.338 
#> GAMLSS-RS iteration 17: Global Deviance = 1580.149 
#> GAMLSS-RS iteration 18: Global Deviance = 1579.983 
#> GAMLSS-RS iteration 19: Global Deviance = 1579.838 
#> GAMLSS-RS iteration 20: Global Deviance = 1579.679 
#> GAMLSS-RS iteration 21: Global Deviance = 1579.534 
#> GAMLSS-RS iteration 22: Global Deviance = 1579.437 
#> 31-08-2020 17:45:59
#> Converged: YES
#> Iterations: 22
#> 
#> Coefficients:
#> (Intercept)	92.0121001471258
#>  log(abs(dist))	-6.09365233585533
#>  b.trans_res	0.205333046120645
#>  p.trans_res	NA
#>  sqrt(b.length)	0.000596462177392219
#>  sqrt(p.length)	-0.00505711406760957
#> 31-08-2020 17:45:59
#> Obtaining normalized randomized quantile residuals for the full dataset...
#> Saving fit to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/fit_5.rds...
#> 31-08-2020 17:45:59
#> Saving all residuals to /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_5.rds...

baits_dir = paste0(base,"/baits")

split_baits_fs(bins_dir,residuals_dir = fits_dir, indices=1:5, output_dir = baits_dir)

#> 31-08-2020 17:45:59
#> Loading bins...
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_1.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_2.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_3.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_4.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/bins/bin_5.rds
#> 31-08-2020 17:45:59
#> Loading residuals...
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_1.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_2.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_3.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_4.rds
#>  /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/fits/residuals_5.rds
#> 31-08-2020 17:45:59
#> Calculating p-values and performing bin-wise FDR-adjustments...
#> 31-08-2020 17:45:59
#> Saving baits...
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_53559.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_71638.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_95958.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_96180.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_102201.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119435.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119522.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_144385.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_162008.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_175487.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_203500.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_243669.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_262362.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_268979.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_272961.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_312780.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_328019.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_341315.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_363262.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_366588.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_466090.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_582536.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_618421.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_658681.rds
#> /tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_661722.rds
#> 31-08-2020 17:45:59
#> Plotting baits...

#> 31-08-2020 17:46:01
#> Done.

#> $output_dir
#> [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits"
#> 
#> $bait_paths
#>  [1] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_53559.rds" 
#>  [2] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_71638.rds" 
#>  [3] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_95958.rds" 
#>  [4] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_96180.rds" 
#>  [5] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_102201.rds"
#>  [6] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119435.rds"
#>  [7] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_119522.rds"
#>  [8] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_144385.rds"
#>  [9] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_162008.rds"
#> [10] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_175487.rds"
#> [11] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_203500.rds"
#> [12] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_243669.rds"
#> [13] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_262362.rds"
#> [14] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_268979.rds"
#> [15] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_272961.rds"
#> [16] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_312780.rds"
#> [17] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_328019.rds"
#> [18] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_341315.rds"
#> [19] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_363262.rds"
#> [20] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_366588.rds"
#> [21] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_466090.rds"
#> [22] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_582536.rds"
#> [23] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_618421.rds"
#> [24] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_658681.rds"
#> [25] "/tmp/RtmpRaB2zI/temp_libpath358f53686a72/peaky/extdata/baits/bait_661722.rds"
#>